whatwg/20111113.html

<!DOCTYPE html>
<title>IRC logs: freenode / #whatwg / 20111113</title>
<style>
ol { font-family: monospace; color: #000; font-size: 90%; list-style: none; margin: 0; padding: 0; }
li { margin: 0; padding: 0; white-space: pre-wrap; }
li:target { background: #f90; }
li.hide { display: none; }
li.flagged { background: yellow; }
li.filtered { color: #aaa; font-size: 60%; }
li.filtered:hover { font-size: 100%; color: #000; height: auto; }
li span { display: inline; margin: 0; padding: 0 10px; background: #eee; color: #000; text-decoration: none; cursor: pointer; }
li.flagged span { background: #f00; }
li:hover span { background: #ff0 }
input#filter { margin: 0 10px; }
label { cursor: pointer; }
p span { background: #ff0; margin-left: 10px; }
</style>
<script src="/irc-logs/s.js"></script>
<h1 id="freenode/whatwg.20111113.log"><a href="/irc-logs/">/irc-logs</a> / freenode / <a href="irc://irc.freenode.net/whatwg">#whatwg</a> / 2011-11-13 / <a href="#end">end</a></h1>
<p>Options: 
<ol id="lines">
 <li id="l-1"><a href="#l-1">#</a> Session Start: Sun Nov 13 00:00:00 2011</li>
 <li id="l-2"><a href="#l-2">#</a> Session Ident: #whatwg</li>
 <li id="l-3"><a href="#l-3">#</a> [00:11] * Quits: Morphous (jan@unaffiliated/amorphous) (Ping timeout: 248 seconds)</li>
 <li id="l-4" _a="heycam"><a href="#l-4">#</a> [00:15] * heycam is now known as heycam|away</li>
 <li id="l-5" _a="heycam|away"><a href="#l-5">#</a> [00:17] * heycam|away is now known as heycam</li>
 <li id="l-6"><a href="#l-6">#</a> [00:20] * Joins: jwalden (~waldo@173-228-26-130.dedicated.static.sonic.net)</li>
 <li id="l-7"><a href="#l-7">#</a> [00:22] * Joins: J_Voracek (~J_Voracek@71.21.195.70)</li>
 <li id="l-8"><a href="#l-8">#</a> [00:27] * Joins: tantek (~tantek@208-90-212-140.PUBLIC.monkeybrains.net)</li>
 <li id="l-9"><a href="#l-9">#</a> [00:27] * Joins: Morphous (jan@unaffiliated/amorphous)</li>
 <li id="l-10"><a href="#l-10">#</a> [00:28] * Quits: J_Voracek (~J_Voracek@71.21.195.70) (Quit: disconnected: Jace Voracek - Jace@Jace-Place.com)</li>
 <li id="l-11"><a href="#l-11">#</a> [00:32] * Joins: Dogan_ (~Dogan@95.15.23.116)</li>
 <li id="l-12"><a href="#l-12">#</a> [00:36] * Quits: sicking (~chatzilla@c-98-210-155-80.hsd1.ca.comcast.net) (Remote host closed the connection)</li>
 <li id="l-13"><a href="#l-13">#</a> [00:38] * Joins: benjoffe_ (~benjoffe_@CPE-121-216-39-241.lnse1.ken.bigpond.net.au)</li>
 <li id="l-14"><a href="#l-14">#</a> [00:43] * Joins: astearns (~anonymous@c-50-132-63-33.hsd1.wa.comcast.net)</li>
 <li id="l-15"><a href="#l-15">#</a> [00:43] * Quits: Telling (~unknown@80-71-135-15.u.parknet.dk) (Quit: ...)</li>
 <li id="l-16"><a href="#l-16">#</a> [00:45] * Quits: stalled (~stalled@unaffiliated/stalled) (Ping timeout: 252 seconds)</li>
 <li id="l-17"><a href="#l-17">#</a> [00:54] * Parts: Dogan_ (~Dogan@95.15.23.116) (&quot;Leaving&quot;)</li>
 <li id="l-18"><a href="#l-18">#</a> [00:54] * Joins: erlehmann (~erlehmann@89.204.137.76)</li>
 <li id="l-19"><a href="#l-19">#</a> [00:55] * Joins: Bass10 (~Bass10@c-76-113-194-7.hsd1.mn.comcast.net)</li>
 <li id="l-20"><a href="#l-20">#</a> [00:56] * Quits: Bass10 (~Bass10@c-76-113-194-7.hsd1.mn.comcast.net) (Max SendQ exceeded)</li>
 <li id="l-21"><a href="#l-21">#</a> [00:57] * Quits: erlehmann (~erlehmann@89.204.137.76) (Client Quit)</li>
 <li id="l-22"><a href="#l-22">#</a> [01:00] * Joins: erlehmann (~erlehmann@89.204.137.76)</li>
 <li id="l-23"><a href="#l-23">#</a> [01:02] * Quits: tantek (~tantek@208-90-212-140.PUBLIC.monkeybrains.net) (Read error: Connection reset by peer)</li>
 <li id="l-24" _a="heycam"><a href="#l-24">#</a> [01:02] * heycam is now known as heycam|away</li>
 <li id="l-25"><a href="#l-25">#</a> [01:03] * Joins: tantek (~tantek@208-90-212-140.PUBLIC.monkeybrains.net)</li>
 <li id="l-26"><a href="#l-26">#</a> [01:07] * Quits: hasather_ (~hasather_@84.38.144.96) (Remote host closed the connection)</li>
 <li id="l-27"><a href="#l-27">#</a> [01:07] * Joins: stalled (~stalled@unaffiliated/stalled)</li>
 <li id="l-28"><a href="#l-28">#</a> [01:08] * Joins: othermaciej (~mjs@c-24-6-209-189.hsd1.ca.comcast.net)</li>
 <li id="l-29"><a href="#l-29">#</a> [01:09] * Joins: saba (~foo@unaffiliated/saba)</li>
 <li id="l-30"><a href="#l-30">#</a> [01:13] * Joins: Telling (~unknown@80-71-135-15.u.parknet.dk)</li>
 <li id="l-31"><a href="#l-31">#</a> [01:14] * Quits: tantek (~tantek@208-90-212-140.PUBLIC.monkeybrains.net) (Read error: Connection reset by peer)</li>
 <li id="l-32"><a href="#l-32">#</a> [01:14] * Joins: tantek (~tantek@208-90-212-140.PUBLIC.monkeybrains.net)</li>
 <li id="l-33"><a href="#l-33">#</a> [01:25] * Quits: erlehmann (~erlehmann@89.204.137.76) (Ping timeout: 245 seconds)</li>
 <li id="l-34"><a href="#l-34">#</a> [01:25] * Joins: rillian_ (~rillian@mist.thaumas.net)</li>
 <li id="l-35"><a href="#l-35">#</a> [01:38] * Joins: erlehmann (~erlehmann@82.113.121.67)</li>
 <li id="l-36"><a href="#l-36">#</a> [01:43] * Quits: necolas (~necolas@5e04326b.bb.sky.com) (Remote host closed the connection)</li>
 <li id="l-37"><a href="#l-37">#</a> [02:00] * Quits: Maurice` (copyman@5ED573FA.cm-7-6b.dynamic.ziggo.nl)</li>
 <li id="l-38"><a href="#l-38">#</a> [02:00] * Quits: rillian_ (~rillian@mist.thaumas.net) (Remote host closed the connection)</li>
 <li id="l-39"><a href="#l-39">#</a> [02:12] * Quits: tantek (~tantek@208-90-212-140.PUBLIC.monkeybrains.net) (Quit: tantek)</li>
 <li id="l-40"><a href="#l-40">#</a> [02:16] * Quits: jacobolus (~jacobolus@c-71-198-169-213.hsd1.ca.comcast.net) (Read error: Operation timed out)</li>
 <li id="l-41"><a href="#l-41">#</a> [02:17] * Joins: jacobolus (~jacobolus@c-71-198-169-213.hsd1.ca.comcast.net)</li>
 <li id="l-42"><a href="#l-42">#</a> [02:17] * Joins: hasather_ (~hasather_@84.38.144.96)</li>
 <li id="l-43"><a href="#l-43">#</a> [02:21] * Quits: Telling (~unknown@80-71-135-15.u.parknet.dk) (Quit: ...)</li>
 <li id="l-44"><a href="#l-44">#</a> [02:22] * Joins: smaug____ (~chatzilla@GGGKMDCCCXL.gprs.sl-laajakaista.fi)</li>
 <li id="l-45"><a href="#l-45">#</a> [02:22] * Quits: hasather_ (~hasather_@84.38.144.96) (Ping timeout: 260 seconds)</li>
 <li id="l-46"><a href="#l-46">#</a> [02:25] * Quits: saba (~foo@unaffiliated/saba) (Ping timeout: 256 seconds)</li>
 <li id="l-47"><a href="#l-47">#</a> [02:25] * Joins: andyg_ (~andyg@geekscape.org)</li>
 <li id="l-48"><a href="#l-48">#</a> [02:26] * Quits: andyg (~andyg@CPE-124-189-148-81.sqcy1.win.bigpond.net.au) (Read error: Connection reset by peer)</li>
 <li id="l-49"><a href="#l-49">#</a> [02:27] * Joins: andyg (~andyg@CPE-124-189-148-81.sqcy1.win.bigpond.net.au)</li>
 <li id="l-50"><a href="#l-50">#</a> [02:28] * Joins: scor (~scor@drupal.org/user/52142/view)</li>
 <li id="l-51"><a href="#l-51">#</a> [02:29] * Quits: andyg_ (~andyg@geekscape.org) (Ping timeout: 255 seconds)</li>
 <li id="l-52"><a href="#l-52">#</a> [02:31] * Quits: scor (~scor@drupal.org/user/52142/view) (Client Quit)</li>
 <li id="l-53"><a href="#l-53">#</a> [02:38] * Quits: connrs (~connrs@conners.plus.com) (Ping timeout: 260 seconds)</li>
 <li id="l-54"><a href="#l-54">#</a> [02:39] * Joins: andyg_ (~andyg@CPE-124-189-148-81.sqcy1.win.bigpond.net.au)</li>
 <li id="l-55"><a href="#l-55">#</a> [02:39] * Quits: andyg (~andyg@CPE-124-189-148-81.sqcy1.win.bigpond.net.au) (Read error: Connection reset by peer)</li>
 <li id="l-56" _a="andyg_"><a href="#l-56">#</a> [02:39] * andyg_ is now known as andyg</li>
 <li id="l-57"><a href="#l-57">#</a> [02:41] * Quits: andyg (~andyg@CPE-124-189-148-81.sqcy1.win.bigpond.net.au) (Read error: Connection reset by peer)</li>
 <li id="l-58"><a href="#l-58">#</a> [02:41] * Joins: andyg (~andyg@CPE-124-189-148-81.sqcy1.win.bigpond.net.au)</li>
 <li id="l-59"><a href="#l-59">#</a> [02:43] * Joins: plutoniix (~plutoniix@ppp-110-169-248-81.revip5.asianet.co.th)</li>
 <li id="l-60"><a href="#l-60">#</a> [02:44] * Joins: connrs (~connrs@conners.plus.com)</li>
 <li id="l-61"><a href="#l-61">#</a> [02:45] * Quits: smaug____ (~chatzilla@GGGKMDCCCXL.gprs.sl-laajakaista.fi) (Ping timeout: 245 seconds)</li>
 <li id="l-62"><a href="#l-62">#</a> [02:47] * Quits: andyg (~andyg@CPE-124-189-148-81.sqcy1.win.bigpond.net.au) (Read error: Connection reset by peer)</li>
 <li id="l-63"><a href="#l-63">#</a> [02:47] * Joins: andyg (~andyg@CPE-124-189-148-81.sqcy1.win.bigpond.net.au)</li>
 <li id="l-64"><a href="#l-64">#</a> [02:57] * Joins: andyg_ (~andyg@CPE-124-189-148-81.sqcy1.win.bigpond.net.au)</li>
 <li id="l-65"><a href="#l-65">#</a> [02:57] * Quits: andyg (~andyg@CPE-124-189-148-81.sqcy1.win.bigpond.net.au) (Read error: Connection reset by peer)</li>
 <li id="l-66" _a="andyg_"><a href="#l-66">#</a> [02:57] * andyg_ is now known as andyg</li>
 <li id="l-67"><a href="#l-67">#</a> [03:04] * Joins: saba (~foo@unaffiliated/saba)</li>
 <li id="l-68"><a href="#l-68">#</a> [03:06] * Joins: andyg_ (~andyg@CPE-124-189-148-81.sqcy1.win.bigpond.net.au)</li>
 <li id="l-69"><a href="#l-69">#</a> [03:08] * Quits: drublic (~drublic@frbg-5f732398.pool.mediaWays.net) (Remote host closed the connection)</li>
 <li id="l-70"><a href="#l-70">#</a> [03:08] * Joins: andyg__ (~andyg@CPE-124-189-148-81.sqcy1.win.bigpond.net.au)</li>
 <li id="l-71"><a href="#l-71">#</a> [03:09] * Quits: andyg (~andyg@CPE-124-189-148-81.sqcy1.win.bigpond.net.au) (Read error: Connection reset by peer)</li>
 <li id="l-72" _a="andyg__"><a href="#l-72">#</a> [03:09] * andyg__ is now known as andyg</li>
 <li id="l-73"><a href="#l-73">#</a> [03:10] * Quits: andyg_ (~andyg@CPE-124-189-148-81.sqcy1.win.bigpond.net.au) (Ping timeout: 245 seconds)</li>
 <li id="l-74"><a href="#l-74">#</a> [03:15] * Joins: dbaron (~dbaron@mail.questnewmarket.co.nz)</li>
 <li id="l-75"><a href="#l-75">#</a> [03:38] * Quits: dbaron (~dbaron@mail.questnewmarket.co.nz) (Quit: 8403864 bytes have been tenured, next gc will be global.)</li>
 <li id="l-76"><a href="#l-76">#</a> [04:12] * Joins: Rik` (~Rik`@88.128.95.130)</li>
 <li id="l-77"><a href="#l-77">#</a> [04:25] * Joins: roc (~chatzilla@121.98.230.221)</li>
 <li id="l-78"><a href="#l-78">#</a> [04:59] * Joins: MikeSmith_ (~MikeSmith@EM1-113-28-101.pool.e-mobile.ne.jp)</li>
 <li id="l-79"><a href="#l-79">#</a> [05:02] * Quits: MikeSmith (~MikeSmith@EM114-48-202-172.pool.e-mobile.ne.jp) (Ping timeout: 248 seconds)</li>
 <li id="l-80" _a="MikeSmith_"><a href="#l-80">#</a> [05:02] * MikeSmith_ is now known as MikeSmith</li>
 <li id="l-81"><a href="#l-81">#</a> [05:04] * Quits: saba (~foo@unaffiliated/saba) (Ping timeout: 240 seconds)</li>
 <li id="l-82"><a href="#l-82">#</a> [05:07] * Joins: tantek (~tantek@208-90-212-140.PUBLIC.monkeybrains.net)</li>
 <li id="l-83"><a href="#l-83">#</a> [05:11] * Quits: tantek (~tantek@208-90-212-140.PUBLIC.monkeybrains.net) (Ping timeout: 240 seconds)</li>
 <li id="l-84"><a href="#l-84">#</a> [05:28] * Joins: nonge__ (~nonge@p5B326AF0.dip.t-dialin.net)</li>
 <li id="l-85"><a href="#l-85">#</a> [05:30] * Joins: dbaron (~dbaron@mail.questnewmarket.co.nz)</li>
 <li id="l-86"><a href="#l-86">#</a> [05:32] * Quits: nonge_ (~nonge@p50829B50.dip.t-dialin.net) (Ping timeout: 258 seconds)</li>
 <li id="l-87"><a href="#l-87">#</a> [06:03] * Joins: tantek (~tantek@208-90-212-140.PUBLIC.monkeybrains.net)</li>
 <li id="l-88"><a href="#l-88">#</a> [06:12] * Quits: roc (~chatzilla@121.98.230.221) (Ping timeout: 260 seconds)</li>
 <li id="l-89"><a href="#l-89">#</a> [06:38] * Joins: benschwarz (~benschwar@59.167.185.148)</li>
 <li id="l-90"><a href="#l-90">#</a> [06:41] * Quits: jwalden (~waldo@173-228-26-130.dedicated.static.sonic.net) (Quit: ChatZilla 0.9.87-3.1450hg.fc15 [XULRunner 7.0.1/20110930134335])</li>
 <li id="l-91"><a href="#l-91">#</a> [06:48] * Joins: rniwa (~rniwa@70-89-66-218-ca.sfba.hfc.comcastbusiness.net)</li>
 <li id="l-92"><a href="#l-92">#</a> [06:49] * Quits: temp02 (~temp01@unaffiliated/temp01) (Read error: No route to host)</li>
 <li id="l-93"><a href="#l-93">#</a> [06:53] * Quits: erlehmann (~erlehmann@82.113.121.67) (Quit: Ex-Chat)</li>
 <li id="l-94"><a href="#l-94">#</a> [06:53] * Joins: temp01 (~temp01@unaffiliated/temp01)</li>
 <li id="l-95" _a="heycam|away"><a href="#l-95">#</a> [06:57] * heycam|away is now known as heycam</li>
 <li id="l-96"><a href="#l-96">#</a> [06:58] * Joins: mpt (~mpt@76.14.69.234)</li>
 <li id="l-97"><a href="#l-97">#</a> [06:58] * Quits: mpt (~mpt@76.14.69.234) (Changing host)</li>
 <li id="l-98"><a href="#l-98">#</a> [06:58] * Joins: mpt (~mpt@canonical/mpt)</li>
 <li id="l-99"><a href="#l-99">#</a> [07:00] * Quits: dbaron (~dbaron@mail.questnewmarket.co.nz) (Quit: 8403864 bytes have been tenured, next gc will be global.)</li>
 <li id="l-100" _a="heycam"><a href="#l-100">#</a> [07:01] * heycam is now known as heycam|away</li>
 <li id="l-101"><a href="#l-101">#</a> [07:03] * Quits: rniwa (~rniwa@70-89-66-218-ca.sfba.hfc.comcastbusiness.net) (Remote host closed the connection)</li>
 <li id="l-102"><a href="#l-102">#</a> [07:03] * Joins: rniwa (~rniwa@216.239.45.130)</li>
 <li id="l-103"><a href="#l-103">#</a> [07:09] * Quits: mpt (~mpt@canonical/mpt) (Ping timeout: 245 seconds)</li>
 <li id="l-104"><a href="#l-104">#</a> [07:37] * Quits: tantek (~tantek@208-90-212-140.PUBLIC.monkeybrains.net) (Quit: tantek)</li>
 <li id="l-105"><a href="#l-105">#</a> [07:57] * Joins: mpt (~mpt@76.14.69.234)</li>
 <li id="l-106"><a href="#l-106">#</a> [07:57] * Quits: mpt (~mpt@76.14.69.234) (Changing host)</li>
 <li id="l-107"><a href="#l-107">#</a> [07:57] * Joins: mpt (~mpt@canonical/mpt)</li>
 <li id="l-108"><a href="#l-108">#</a> [07:58] * Quits: Rik` (~Rik`@88.128.95.130) (Ping timeout: 240 seconds)</li>
 <li id="l-109"><a href="#l-109">#</a> [08:12] * Joins: GlitchMr (~glitchmr@178-36-13-251.adsl.inetia.pl)</li>
 <li id="l-110"><a href="#l-110">#</a> [08:26] * Joins: ezoe (~ezoe@112-68-245-227f1.kyt1.eonet.ne.jp)</li>
 <li id="l-111"><a href="#l-111">#</a> [08:35] * Quits: mpt (~mpt@canonical/mpt) (Read error: Operation timed out)</li>
 <li id="l-112"><a href="#l-112">#</a> [09:00] * Quits: dydx (~dydz@adsl-75-36-190-15.dsl.pltn13.sbcglobal.net) (Quit: dydx)</li>
 <li id="l-113"><a href="#l-113">#</a> [09:09] * Joins: dydx (~dydz@adsl-75-36-190-15.dsl.pltn13.sbcglobal.net)</li>
 <li id="l-114"><a href="#l-114">#</a> [09:13] * Quits: dydx (~dydz@adsl-75-36-190-15.dsl.pltn13.sbcglobal.net) (Ping timeout: 260 seconds)</li>
 <li id="l-115"><a href="#l-115">#</a> [09:19] * Joins: Rik` (~Rik`@p5DDBB05D.dip.t-dialin.net)</li>
 <li id="l-116"><a href="#l-116">#</a> [09:21] * Joins: Areks (~Areks@95-26-164-46.broadband.corbina.ru)</li>
 <li id="l-117"><a href="#l-117">#</a> [09:23] * Joins: mpt (~mpt@76.14.69.234)</li>
 <li id="l-118"><a href="#l-118">#</a> [09:23] * Quits: mpt (~mpt@76.14.69.234) (Changing host)</li>
 <li id="l-119"><a href="#l-119">#</a> [09:23] * Joins: mpt (~mpt@canonical/mpt)</li>
 <li id="l-120"><a href="#l-120">#</a> [09:25] * Quits: Rik` (~Rik`@p5DDBB05D.dip.t-dialin.net) (Remote host closed the connection)</li>
 <li id="l-121"><a href="#l-121">#</a> [09:26] * Quits: Areks (~Areks@95-26-164-46.broadband.corbina.ru) (Ping timeout: 258 seconds)</li>
 <li id="l-122"><a href="#l-122">#</a> [09:34] * Quits: mpt (~mpt@canonical/mpt) (Ping timeout: 255 seconds)</li>
 <li id="l-123"><a href="#l-123">#</a> [09:35] * Quits: ezoe (~ezoe@112-68-245-227f1.kyt1.eonet.ne.jp) (Quit: And Now for Something Completely Different.)</li>
 <li id="l-124"><a href="#l-124">#</a> [09:48] * Joins: FlorianX (~Dimitri@p4FCF7A7E.dip.t-dialin.net)</li>
 <li id="l-125"><a href="#l-125">#</a> [09:49] * Quits: Druid_ (~Druid@p5B13742F.dip.t-dialin.net) (Ping timeout: 260 seconds)</li>
 <li id="l-126"><a href="#l-126">#</a> [09:53] * Joins: dbaron (~dbaron@mail.questnewmarket.co.nz)</li>
 <li id="l-127"><a href="#l-127">#</a> [09:54] * Joins: Druid_ (~Druid@p5B13711F.dip.t-dialin.net)</li>
 <li id="l-128"><a href="#l-128">#</a> [10:01] * Quits: rniwa (~rniwa@216.239.45.130) (Quit: rniwa)</li>
 <li id="l-129"><a href="#l-129">#</a> [10:03] * Joins: dydx (~dydz@adsl-75-36-190-15.dsl.pltn13.sbcglobal.net)</li>
 <li id="l-130"><a href="#l-130">#</a> [10:07] * Quits: dydx (~dydz@adsl-75-36-190-15.dsl.pltn13.sbcglobal.net) (Ping timeout: 260 seconds)</li>
 <li id="l-131"><a href="#l-131">#</a> [10:11] * Quits: dbaron (~dbaron@mail.questnewmarket.co.nz) (Quit: g'night)</li>
 <li id="l-132"><a href="#l-132">#</a> [10:20] * Joins: danielfilho_ (~daniel@187.31.77.7)</li>
 <li id="l-133"><a href="#l-133">#</a> [10:21] * Quits: danielfilho (~daniel@187.31.77.7) (Read error: Operation timed out)</li>
 <li id="l-134" _a="danielfilho_"><a href="#l-134">#</a> [10:21] * danielfilho_ is now known as danielfilho</li>
 <li id="l-135"><a href="#l-135">#</a> [10:39] * Joins: hasather_ (~hasather_@84.38.144.96)</li>
 <li id="l-136"><a href="#l-136">#</a> [10:52] * Quits: benschwarz (~benschwar@59.167.185.148) (Quit: Linkinus - <a href="http://linkinus.com">http://linkinus.com</a>)</li>
 <li id="l-137"><a href="#l-137">#</a> [10:53] * Quits: espadrine (~thaddee_t@acces2064.res.insa-lyon.fr) (Ping timeout: 276 seconds)</li>
 <li id="l-138"><a href="#l-138">#</a> [10:57] * Joins: dydx (~dydz@adsl-75-36-190-15.dsl.pltn13.sbcglobal.net)</li>
 <li id="l-139"><a href="#l-139">#</a> [11:00] * Joins: MikeSmith_ (~MikeSmith@EM1-113-48-31.pool.e-mobile.ne.jp)</li>
 <li id="l-140"><a href="#l-140">#</a> [11:01] * Quits: dydx (~dydz@adsl-75-36-190-15.dsl.pltn13.sbcglobal.net) (Ping timeout: 240 seconds)</li>
 <li id="l-141"><a href="#l-141">#</a> [11:02] * Quits: MikeSmith (~MikeSmith@EM1-113-28-101.pool.e-mobile.ne.jp) (Ping timeout: 255 seconds)</li>
 <li id="l-142" _a="MikeSmith_"><a href="#l-142">#</a> [11:02] * MikeSmith_ is now known as MikeSmith</li>
 <li id="l-143"><a href="#l-143">#</a> [11:06] * Joins: othermaciej_ (~mjs@c-24-6-209-189.hsd1.ca.comcast.net)</li>
 <li id="l-144"><a href="#l-144">#</a> [11:06] * Quits: othermaciej (~mjs@c-24-6-209-189.hsd1.ca.comcast.net) (Read error: Connection reset by peer)</li>
 <li id="l-145" _a="othermaciej_"><a href="#l-145">#</a> [11:06] * othermaciej_ is now known as othermaciej</li>
 <li id="l-146"><a href="#l-146">#</a> [11:06] * Quits: othermaciej (~mjs@c-24-6-209-189.hsd1.ca.comcast.net) (Client Quit)</li>
 <li id="l-147"><a href="#l-147">#</a> [11:13] * Joins: espadrine (~thaddee_t@acces2064.res.insa-lyon.fr)</li>
 <li id="l-148"><a href="#l-148">#</a> [11:14] * Quits: andyg (~andyg@CPE-124-189-148-81.sqcy1.win.bigpond.net.au) (Quit: andyg)</li>
 <li id="l-149"><a href="#l-149">#</a> [11:20] * Joins: Maurice (copyman@5ED573FA.cm-7-6b.dynamic.ziggo.nl)</li>
 <li id="l-150"><a href="#l-150">#</a> [11:33] * Joins: Ms2ger (~Ms2ger@91.181.168.54)</li>
 <li id="l-151"><a href="#l-151">#</a> [11:39] * Joins: tomasf (~tom@c-b7dbe555.024-204-6c6b7012.cust.bredbandsbolaget.se)</li>
 <li id="l-152"><a href="#l-152">#</a> [11:41] * Joins: andyg (~andyg@CPE-124-189-148-81.sqcy1.win.bigpond.net.au)</li>
 <li id="l-153"><a href="#l-153">#</a> [11:42] * Joins: svl (~me@ip565744a7.direct-adsl.nl)</li>
 <li id="l-154"><a href="#l-154">#</a> [11:43] * Joins: saba (~foo@unaffiliated/saba)</li>
 <li id="l-155"><a href="#l-155">#</a> [11:44] * Joins: jarek (~jarek@bdc167.neoplus.adsl.tpnet.pl)</li>
 <li id="l-156"><a href="#l-156">#</a> [11:44] * Quits: jarek (~jarek@bdc167.neoplus.adsl.tpnet.pl) (Changing host)</li>
 <li id="l-157"><a href="#l-157">#</a> [11:44] * Joins: jarek (~jarek@unaffiliated/jarek)</li>
 <li id="l-158"><a href="#l-158">#</a> [12:12] * Quits: shepazu (~shepazu@108-70-132-46.lightspeed.rlghnc.sbcglobal.net) (Quit: shepazu)</li>
 <li id="l-159"><a href="#l-159">#</a> [12:24] * Joins: erlehmann (~erlehmann@82.113.121.67)</li>
 <li id="l-160"><a href="#l-160">#</a> [12:41] * Joins: drublic (~drublic@frbg-5d84ed6c.pool.mediaWays.net)</li>
 <li id="l-161"><a href="#l-161">#</a> [12:52] * Joins: smaug____ (~chatzilla@GGGKMMCMXCIV.gprs.sl-laajakaista.fi)</li>
 <li id="l-162"><a href="#l-162">#</a> [12:58] * Joins: Telling (~unknown@80-71-135-15.u.parknet.dk)</li>
 <li id="l-163"><a href="#l-163">#</a> [13:05] * Joins: PalleZingmark (~Adium@5e0571a8.bb.sky.com)</li>
 <li id="l-164"><a href="#l-164">#</a> [13:25] * Quits: JuanDaugherty (~Ren@cpe-72-228-177-92.buffalo.res.rr.com) (Ping timeout: 258 seconds)</li>
 <li id="l-165"><a href="#l-165">#</a> [13:28] * Quits: GlitchMr (~glitchmr@178-36-13-251.adsl.inetia.pl) (Quit: intentional quit - disconnect)</li>
 <li id="l-166"><a href="#l-166">#</a> [13:31] * Joins: GlitchMr (~glitchmr@178-36-13-251.adsl.inetia.pl)</li>
 <li id="l-167"><a href="#l-167">#</a> [13:31] * Quits: andyg (~andyg@CPE-124-189-148-81.sqcy1.win.bigpond.net.au) (Quit: andyg)</li>
 <li id="l-168"><a href="#l-168">#</a> [13:33] * Quits: smaug____ (~chatzilla@GGGKMMCMXCIV.gprs.sl-laajakaista.fi) (Ping timeout: 248 seconds)</li>
 <li id="l-169"><a href="#l-169">#</a> [13:43] * Quits: Druid_ (~Druid@p5B13711F.dip.t-dialin.net) (Ping timeout: 260 seconds)</li>
 <li id="l-170"><a href="#l-170">#</a> [13:48] * Joins: Druid_ (~Druid@p5B1376A3.dip.t-dialin.net)</li>
 <li id="l-171"><a href="#l-171">#</a> [13:50] * Joins: JuanDaugherty (~Ren@cpe-72-228-177-92.buffalo.res.rr.com)</li>
 <li id="l-172"><a href="#l-172">#</a> [13:52] * Quits: hasather_ (~hasather_@84.38.144.96) (Remote host closed the connection)</li>
 <li id="l-173"><a href="#l-173">#</a> [13:57] * Joins: andyg (~andyg@CPE-124-189-148-81.sqcy1.win.bigpond.net.au)</li>
 <li id="l-174"><a href="#l-174">#</a> [14:10] * Quits: GlitchMr (~glitchmr@178-36-13-251.adsl.inetia.pl) (Quit: GlitchMr)</li>
 <li id="l-175"><a href="#l-175">#</a> [14:13] * Joins: tndH (~Rob@cpc16-seac19-2-0-cust234.7-2.cable.virginmedia.com)</li>
 <li id="l-176"><a href="#l-176">#</a> [14:24] * Quits: tomaw (tom@freenode/staff/tomaw) (Quit: Quitting)</li>
 <li id="l-177"><a href="#l-177">#</a> [14:25] * Joins: tomaw (tom@freenode/staff/tomaw)</li>
 <li id="l-178"><a href="#l-178">#</a> [14:27] * Quits: tomaw (tom@freenode/staff/tomaw) (Quit: Quitting)</li>
 <li id="l-179"><a href="#l-179">#</a> [14:27] * Joins: tomaw (tom@freenode/staff/tomaw)</li>
 <li id="l-180"><a href="#l-180">#</a> [14:29] * Quits: connrs (~connrs@conners.plus.com) (*.net *.split)</li>
 <li id="l-181"><a href="#l-181">#</a> [14:29] * Quits: shetech (~shetech@c-76-126-167-49.hsd1.ca.comcast.net) (*.net *.split)</li>
 <li id="l-182"><a href="#l-182">#</a> [14:29] * Quits: _bga (~bga@95-55-61-178.dynamic.avangarddsl.ru) (*.net *.split)</li>
 <li id="l-183"><a href="#l-183">#</a> [14:29] * Quits: krijn (u2319@gateway/web/irccloud.com/x-gdyudgtmfzwgpoei) (*.net *.split)</li>
 <li id="l-184"><a href="#l-184">#</a> [14:29] * Quits: lensco (u4054@gateway/web/irccloud.com/x-smwpdzgtvjkkbdni) (*.net *.split)</li>
 <li id="l-185"><a href="#l-185">#</a> [14:29] * Quits: antti_s (~asal@backport.reaktor.fi) (*.net *.split)</li>
 <li id="l-186"><a href="#l-186">#</a> [14:29] * Quits: stalled (~stalled@unaffiliated/stalled) (*.net *.split)</li>
 <li id="l-187"><a href="#l-187">#</a> [14:29] * Quits: hober (~ted@unaffiliated/hober) (*.net *.split)</li>
 <li id="l-188"><a href="#l-188">#</a> [14:29] * Quits: webben (~benjamin@173-203-84-17.static.cloud-ips.com) (*.net *.split)</li>
 <li id="l-189"><a href="#l-189">#</a> [14:29] * Quits: JuanDaugherty (~Ren@cpe-72-228-177-92.buffalo.res.rr.com) (*.net *.split)</li>
 <li id="l-190"><a href="#l-190">#</a> [14:29] * Quits: svl (~me@ip565744a7.direct-adsl.nl) (*.net *.split)</li>
 <li id="l-191"><a href="#l-191">#</a> [14:29] * Quits: astearns (~anonymous@c-50-132-63-33.hsd1.wa.comcast.net) (*.net *.split)</li>
 <li id="l-192"><a href="#l-192">#</a> [14:29] * Quits: bzed (~bzed@devel.recluse.de) (*.net *.split)</li>
 <li id="l-193"><a href="#l-193">#</a> [14:29] * Quits: bobylito (u3929@gateway/web/irccloud.com/x-cknembtuoifrrjlj) (*.net *.split)</li>
 <li id="l-194"><a href="#l-194">#</a> [14:29] * Quits: heycam|away (~cam@wok.mcc.id.au) (*.net *.split)</li>
 <li id="l-195"><a href="#l-195">#</a> [14:29] * Quits: wirepair (fbi@random.supermario.org) (*.net *.split)</li>
 <li id="l-196"><a href="#l-196">#</a> [14:29] * Quits: ivan`` (~ivan@li125-242.members.linode.com) (*.net *.split)</li>
 <li id="l-197"><a href="#l-197">#</a> [14:29] * Quits: hasather (~davidh@pat-tdc.opera.com) (*.net *.split)</li>
 <li id="l-198"><a href="#l-198">#</a> [14:29] * Quits: mven (~mven__@169.241.49.57) (*.net *.split)</li>
 <li id="l-199"><a href="#l-199">#</a> [14:29] * Quits: llrcombs (~llrcombs@64.130.210.214) (*.net *.split)</li>
 <li id="l-200"><a href="#l-200">#</a> [14:29] * Quits: purefusion (~purefusio@unaffiliated/purefusion) (*.net *.split)</li>
 <li id="l-201"><a href="#l-201">#</a> [14:29] * Quits: danielfilho (~daniel@187.31.77.7) (*.net *.split)</li>
 <li id="l-202"><a href="#l-202">#</a> [14:29] * Quits: _salavas (~salavas@c83-248-102-83.bredband.comhem.se) (*.net *.split)</li>
 <li id="l-203"><a href="#l-203">#</a> [14:29] * Quits: payman (~payman@pat.se.opera.com) (*.net *.split)</li>
 <li id="l-204"><a href="#l-204">#</a> [14:29] * Quits: jennb (~jennb@74.125.59.65) (*.net *.split)</li>
 <li id="l-205"><a href="#l-205">#</a> [14:29] * Quits: zewt (~x@c-24-62-196-44.hsd1.ma.comcast.net) (*.net *.split)</li>
 <li id="l-206"><a href="#l-206">#</a> [14:29] * Quits: franksalim (~frank@64-71-23-251.static.wiline.com) (*.net *.split)</li>
 <li id="l-207"><a href="#l-207">#</a> [14:29] * Quits: eric_carlson (~ericc@adsl-67-112-12-110.dsl.anhm01.pacbell.net) (*.net *.split)</li>
 <li id="l-208"><a href="#l-208">#</a> [14:29] * Quits: annevk5 (u2483@gateway/web/irccloud.com/x-mtcimozrgjewgrki) (*.net *.split)</li>
 <li id="l-209"><a href="#l-209">#</a> [14:29] * Quits: manu-db (~msporny@digitalbazaar.com) (*.net *.split)</li>
 <li id="l-210"><a href="#l-210">#</a> [14:29] * Quits: Hixie (~ianh@trivini.no) (*.net *.split)</li>
 <li id="l-211"><a href="#l-211">#</a> [14:29] * Quits: toyoshim_ (~toyoshim@yuri.twintail.org) (*.net *.split)</li>
 <li id="l-212"><a href="#l-212">#</a> [14:29] * Quits: paul_irish (~paul_iris@ve.hsh6wjwx.vesrv.com) (*.net *.split)</li>
 <li id="l-213"><a href="#l-213">#</a> [14:29] * Quits: GPH-Zeke (~GPHemsley@pdpc/supporter/student/GPHemsley) (*.net *.split)</li>
 <li id="l-214"><a href="#l-214">#</a> [14:29] * Quits: Moo_O (~quassel@herd37.twinapex.fi) (*.net *.split)</li>
 <li id="l-215"><a href="#l-215">#</a> [14:29] * Quits: Telling (~unknown@80-71-135-15.u.parknet.dk) (*.net *.split)</li>
 <li id="l-216"><a href="#l-216">#</a> [14:29] * Quits: saba (~foo@unaffiliated/saba) (*.net *.split)</li>
 <li id="l-217"><a href="#l-217">#</a> [14:29] * Quits: foolip_ (u3586@gateway/web/irccloud.com/x-cjhbzbmtwahekqba) (*.net *.split)</li>
 <li id="l-218"><a href="#l-218">#</a> [14:29] * Quits: kbrosnan (~kbrosnan@firefox/community/qa/kbrosnan) (*.net *.split)</li>
 <li id="l-219"><a href="#l-219">#</a> [14:29] * Quits: doublec (~doublec@unaffiliated/doublec) (*.net *.split)</li>
 <li id="l-220"><a href="#l-220">#</a> [14:29] * Quits: jarib (~jarib@unaffiliated/jarib) (*.net *.split)</li>
 <li id="l-221"><a href="#l-221">#</a> [14:29] * Quits: hoodow (~hoodow@pdpc/supporter/active/hoodow) (*.net *.split)</li>
 <li id="l-222"><a href="#l-222">#</a> [14:29] * Quits: nunnun (~nunnun@2001:268:355:1:20c:29ff:fed5:973a) (*.net *.split)</li>
 <li id="l-223"><a href="#l-223">#</a> [14:29] * Quits: drublic (~drublic@frbg-5d84ed6c.pool.mediaWays.net) (*.net *.split)</li>
 <li id="l-224"><a href="#l-224">#</a> [14:29] * Quits: benjoffe_ (~benjoffe_@CPE-121-216-39-241.lnse1.ken.bigpond.net.au) (*.net *.split)</li>
 <li id="l-225"><a href="#l-225">#</a> [14:29] * Quits: Peter` (~peter@nishino.lvp-media.com) (*.net *.split)</li>
 <li id="l-226"><a href="#l-226">#</a> [14:29] * Quits: jernoble|afk (~jernoble@17.212.152.13) (*.net *.split)</li>
 <li id="l-227"><a href="#l-227">#</a> [14:29] * Quits: Yudai (~Yudai@p656629.tokynt01.ap.so-net.ne.jp) (*.net *.split)</li>
 <li id="l-228"><a href="#l-228">#</a> [14:29] * Quits: beowulf (u116@pdpc/supporter/professional/beowulf) (*.net *.split)</li>
 <li id="l-229"><a href="#l-229">#</a> [14:29] * Quits: niftylettuce (u2733@gateway/web/irccloud.com/x-ayzqrjvdyvlvqbpy) (*.net *.split)</li>
 <li id="l-230"><a href="#l-230">#</a> [14:29] * Quits: Velmont (odinho@knuth.ping.uio.no) (*.net *.split)</li>
 <li id="l-231"><a href="#l-231">#</a> [14:29] * Quits: hendry (~hendry@85.119.83.83) (*.net *.split)</li>
 <li id="l-232"><a href="#l-232">#</a> [14:29] * Quits: ben_alman_ (~ben_alman@web126.webfaction.com) (*.net *.split)</li>
 <li id="l-233"><a href="#l-233">#</a> [14:29] * Quits: ParadoX- (parad0x@hades.spexhost.com) (*.net *.split)</li>
 <li id="l-234"><a href="#l-234">#</a> [14:29] * Quits: gavin (~gavin@firefox/developer/gavin) (*.net *.split)</li>
 <li id="l-235"><a href="#l-235">#</a> [14:29] * Quits: bentruyman (~bentruyma@li159-104.members.linode.com) (*.net *.split)</li>
 <li id="l-236"><a href="#l-236">#</a> [14:29] * Quits: silky (~silky@pool-74-108-142-22.nycmny.fios.verizon.net) (*.net *.split)</li>
 <li id="l-237"><a href="#l-237">#</a> [14:29] * Quits: danheberden (~danheberd@li225-35.members.linode.com) (*.net *.split)</li>
 <li id="l-238"><a href="#l-238">#</a> [14:29] * Quits: Kingdutch (~kingdutch@188.200.149.217) (*.net *.split)</li>
 <li id="l-239"><a href="#l-239">#</a> [14:29] * Quits: plutoniix (~plutoniix@ppp-110-169-248-81.revip5.asianet.co.th) (*.net *.split)</li>
 <li id="l-240"><a href="#l-240">#</a> [14:30] * Quits: Phae (u455@gateway/web/irccloud.com/x-kjkaumlrcwwsuwop) (*.net *.split)</li>
 <li id="l-241"><a href="#l-241">#</a> [14:30] * Quits: gavin_ (~gavin@76.14.70.183) (*.net *.split)</li>
 <li id="l-242"><a href="#l-242">#</a> [14:30] * Quits: Philip` (~philip@92.243.11.39) (*.net *.split)</li>
 <li id="l-243"><a href="#l-243">#</a> [14:30] * Quits: TabAtkins (~tabatkins@216.239.45.4) (*.net *.split)</li>
 <li id="l-244"><a href="#l-244">#</a> [14:30] * Quits: inimino (~inimino@69.164.198.129) (*.net *.split)</li>
 <li id="l-245"><a href="#l-245">#</a> [14:30] * Quits: matjas (u2247@gateway/web/irccloud.com/x-wrsfupcozgamgysp) (*.net *.split)</li>
 <li id="l-246"><a href="#l-246">#</a> [14:30] * Quits: hsivonen (~hsivonen@srv-e205.esp.mediateam.fi) (*.net *.split)</li>
 <li id="l-247"><a href="#l-247">#</a> [14:30] * Quits: reggna (~reggna@godis.olf.sgsnet.se) (*.net *.split)</li>
 <li id="l-248"><a href="#l-248">#</a> [14:30] * Quits: Maurice (copyman@5ED573FA.cm-7-6b.dynamic.ziggo.nl) (*.net *.split)</li>
 <li id="l-249"><a href="#l-249">#</a> [14:30] * Quits: dglazkov (u4270@gateway/web/irccloud.com/x-eteinghaavexdwwy) (*.net *.split)</li>
 <li id="l-250"><a href="#l-250">#</a> [14:30] * Quits: FireFly (firefly@unaffiliated/firefly) (*.net *.split)</li>
 <li id="l-251"><a href="#l-251">#</a> [14:30] * Quits: crash\ (bouncer@lubyte.de) (*.net *.split)</li>
 <li id="l-252"><a href="#l-252">#</a> [14:30] * Quits: Workshiva (~Dashiva@74.125.57.33) (*.net *.split)</li>
 <li id="l-253"><a href="#l-253">#</a> [14:30] * Quits: MikeSmith (~MikeSmith@EM1-113-48-31.pool.e-mobile.ne.jp) (*.net *.split)</li>
 <li id="l-254"><a href="#l-254">#</a> [14:30] * Quits: nonge__ (~nonge@p5B326AF0.dip.t-dialin.net) (*.net *.split)</li>
 <li id="l-255"><a href="#l-255">#</a> [14:30] * Quits: jacobolus (~jacobolus@c-71-198-169-213.hsd1.ca.comcast.net) (*.net *.split)</li>
 <li id="l-256"><a href="#l-256">#</a> [14:30] * Quits: PrgmrBill (~PrgmrBill@unaffiliated/prgmrbill) (*.net *.split)</li>
 <li id="l-257"><a href="#l-257">#</a> [14:30] * Quits: gwillen (~gwillen@unaffiliated/gwillen) (*.net *.split)</li>
 <li id="l-258"><a href="#l-258">#</a> [14:30] * Quits: Russ_ (~quassel@nat/transgaming/x-llrapuqqiqvpahbf) (*.net *.split)</li>
 <li id="l-259"><a href="#l-259">#</a> [14:30] * Quits: Necrathex (~nectop@82-170-160-25.ip.telfort.nl) (*.net *.split)</li>
 <li id="l-260"><a href="#l-260">#</a> [14:30] * Quits: yutak (~yutak@2401:fa00:4:1004:baac:6fff:fe99:adfb) (*.net *.split)</li>
 <li id="l-261"><a href="#l-261">#</a> [14:30] * Quits: Rubennn (~quassel@2a02:348:33:5823::1) (*.net *.split)</li>
 <li id="l-262"><a href="#l-262">#</a> [14:30] * Quits: michaelw (~michaelw@2a01:4f8:131:23c1::5) (*.net *.split)</li>
 <li id="l-263"><a href="#l-263">#</a> [14:30] * Quits: tomasf (~tom@c-b7dbe555.024-204-6c6b7012.cust.bredbandsbolaget.se) (*.net *.split)</li>
 <li id="l-264"><a href="#l-264">#</a> [14:30] * Quits: gkellogg (~gregg@c-98-248-150-91.hsd1.ca.comcast.net) (*.net *.split)</li>
 <li id="l-265"><a href="#l-265">#</a> [14:30] * Quits: arv (u4269@gateway/web/irccloud.com/x-hjpsmrvjscknpads) (*.net *.split)</li>
 <li id="l-266"><a href="#l-266">#</a> [14:30] * Quits: boblet (u1921@gateway/web/irccloud.com/x-rckvknshduohrcqr) (*.net *.split)</li>
 <li id="l-267"><a href="#l-267">#</a> [14:30] * Quits: Obvious_MkII (tachikoma@188.226.74.2) (*.net *.split)</li>
 <li id="l-268"><a href="#l-268">#</a> [14:30] * Quits: rektide (~rektide@67.217.170.145) (*.net *.split)</li>
 <li id="l-269"><a href="#l-269">#</a> [14:30] * Quits: wilhelm (~wilhelm@178.255.149.98) (*.net *.split)</li>
 <li id="l-270"><a href="#l-270">#</a> [14:30] * Quits: nw` (eero@heaven.unlink.org) (*.net *.split)</li>
 <li id="l-271"><a href="#l-271">#</a> [14:30] * Quits: Dashiva (Dashiva@wikia/Dashiva) (*.net *.split)</li>
 <li id="l-272"><a href="#l-272">#</a> [14:30] * Quits: MrOpposite (~MrOpposit@unaffiliated/mropposite) (*.net *.split)</li>
 <li id="l-273"><a href="#l-273">#</a> [14:30] * Quits: djr (~djr@208-75-85-139.slicehost.net) (*.net *.split)</li>
 <li id="l-274"><a href="#l-274">#</a> [14:30] * Quits: tndH (~Rob@cpc16-seac19-2-0-cust234.7-2.cable.virginmedia.com) (*.net *.split)</li>
 <li id="l-275"><a href="#l-275">#</a> [14:30] * Quits: andyg (~andyg@CPE-124-189-148-81.sqcy1.win.bigpond.net.au) (*.net *.split)</li>
 <li id="l-276"><a href="#l-276">#</a> [14:30] * Quits: erlehmann (~erlehmann@82.113.121.67) (*.net *.split)</li>
 <li id="l-277"><a href="#l-277">#</a> [14:30] * Quits: jarek (~jarek@unaffiliated/jarek) (*.net *.split)</li>
 <li id="l-278"><a href="#l-278">#</a> [14:30] * Quits: Ms2ger (~Ms2ger@91.181.168.54) (*.net *.split)</li>
 <li id="l-279"><a href="#l-279">#</a> [14:30] * Quits: temp01 (~temp01@unaffiliated/temp01) (*.net *.split)</li>
 <li id="l-280"><a href="#l-280">#</a> [14:30] * Quits: Morphous (jan@unaffiliated/amorphous) (*.net *.split)</li>
 <li id="l-281"><a href="#l-281">#</a> [14:30] * Quits: kennyluck (~kennyluck@114-43-120-164.dynamic.hinet.net) (*.net *.split)</li>
 <li id="l-282"><a href="#l-282">#</a> [14:30] * Quits: AryehGregor (~Simetrica@mediawiki/simetrical) (*.net *.split)</li>
 <li id="l-283"><a href="#l-283">#</a> [14:30] * Quits: romainhuet (u2533@gateway/web/irccloud.com/x-mdxlmkhggzrtgfrt) (*.net *.split)</li>
 <li id="l-284"><a href="#l-284">#</a> [14:30] * Quits: jeremyselier (u2513@gateway/web/irccloud.com/x-mboguljwqhlljoym) (*.net *.split)</li>
 <li id="l-285"><a href="#l-285">#</a> [14:30] * Quits: wycats (u79@gateway/web/irccloud.com/x-uhodpmtelvvwclku) (*.net *.split)</li>
 <li id="l-286"><a href="#l-286">#</a> [14:30] * Quits: eae (u4278@gateway/web/irccloud.com/x-ptpufxwfrnbvkwnn) (*.net *.split)</li>
 <li id="l-287"><a href="#l-287">#</a> [14:30] * Quits: kinetik (~kinetik@121.98.132.55) (*.net *.split)</li>
 <li id="l-288"><a href="#l-288">#</a> [14:30] * Quits: michel_v (~tofu@unaffiliated/michelv/x-000000001) (*.net *.split)</li>
 <li id="l-289"><a href="#l-289">#</a> [14:30] * Quits: manu` (~chatzilla@pool-74-107-162-72.ronkva.east.verizon.net) (*.net *.split)</li>
 <li id="l-290"><a href="#l-290">#</a> [14:30] * Quits: matijsb (u2278@gateway/web/irccloud.com/x-abqdczhjtjictiww) (*.net *.split)</li>
 <li id="l-291"><a href="#l-291">#</a> [14:30] * Quits: Jedi_ (~Jedi@jedi.org) (*.net *.split)</li>
 <li id="l-292"><a href="#l-292">#</a> [14:30] * Quits: jcranmer (~jcranmer@ltsp2.csl.tjhsst.edu) (*.net *.split)</li>
 <li id="l-293"><a href="#l-293">#</a> [14:30] * Quits: ukai (ukai@nat/google/x-sufkgyystqmagdvi) (*.net *.split)</li>
 <li id="l-294"><a href="#l-294">#</a> [14:30] * Quits: Druid_ (~Druid@p5B1376A3.dip.t-dialin.net) (*.net *.split)</li>
 <li id="l-295"><a href="#l-295">#</a> [14:30] * Quits: FlorianX (~Dimitri@p4FCF7A7E.dip.t-dialin.net) (*.net *.split)</li>
 <li id="l-296"><a href="#l-296">#</a> [14:30] * Quits: esc__ (~esc-ape@193.170.99.241) (*.net *.split)</li>
 <li id="l-297"><a href="#l-297">#</a> [14:30] * Quits: asmodai (asmodai@178-85-121-247.dynamic.upc.nl) (*.net *.split)</li>
 <li id="l-298"><a href="#l-298">#</a> [14:30] * Quits: Martijnc (~Martijn@d54C02C64.access.telenet.be) (*.net *.split)</li>
 <li id="l-299"><a href="#l-299">#</a> [14:30] * Quits: Mike[tm] (~MikeSmith@sideshowbarker.net) (*.net *.split)</li>
 <li id="l-300"><a href="#l-300">#</a> [14:30] * Quits: tellnes (~KrooniX@ec2-79-125-26-36.eu-west-1.compute.amazonaws.com) (*.net *.split)</li>
 <li id="l-301"><a href="#l-301">#</a> [14:30] * Quits: broquaint (a52b469f90@78.47.79.137) (*.net *.split)</li>
 <li id="l-302"><a href="#l-302">#</a> [14:30] * Quits: cgcardona (~cgcardona@unaffiliated/cgcardona) (*.net *.split)</li>
 <li id="l-303"><a href="#l-303">#</a> [14:30] * Quits: bensmithett (~bensmithe@115.146.71.1) (*.net *.split)</li>
 <li id="l-304"><a href="#l-304">#</a> [14:30] * Quits: pererik (~pe@unaffiliated/pererik) (*.net *.split)</li>
 <li id="l-305"><a href="#l-305">#</a> [14:30] * Quits: lhnz (~lhnz@188-223-83-48.zone14.bethere.co.uk) (*.net *.split)</li>
 <li id="l-306"><a href="#l-306">#</a> [14:30] * Quits: Taggnostr (~quassel@dyn57-215.yok.fi) (*.net *.split)</li>
 <li id="l-307"><a href="#l-307">#</a> [14:30] * Quits: nlogax (~nlogax@unaffiliated/nlogax) (*.net *.split)</li>
 <li id="l-308"><a href="#l-308">#</a> [14:30] * Quits: PalleZingmark (~Adium@5e0571a8.bb.sky.com) (*.net *.split)</li>
 <li id="l-309"><a href="#l-309">#</a> [14:30] * Quits: cpearce (~chatzilla@ip-118-90-36-154.xdsl.xnet.co.nz) (*.net *.split)</li>
 <li id="l-310"><a href="#l-310">#</a> [14:30] * Quits: jgraham (~jgraham@74.53.238.210) (*.net *.split)</li>
 <li id="l-311"><a href="#l-311">#</a> [14:30] * Quits: annevk (~annevk@5355737B.cm-6-6b.dynamic.ziggo.nl) (*.net *.split)</li>
 <li id="l-312"><a href="#l-312">#</a> [14:30] * Quits: riven (~riven@pdpc/supporter/professional/riven) (*.net *.split)</li>
 <li id="l-313"><a href="#l-313">#</a> [14:30] * Quits: slartsa (~slartsa@alpha.pumppumedia.com) (*.net *.split)</li>
 <li id="l-314"><a href="#l-314">#</a> [14:30] * Quits: foolip (~philip@83.218.67.122) (*.net *.split)</li>
 <li id="l-315"><a href="#l-315">#</a> [14:30] * Quits: tmzt (~tmzt@adsl-99-164-105-221.dsl.akrnoh.sbcglobal.net) (*.net *.split)</li>
 <li id="l-316"><a href="#l-316">#</a> [14:30] * Quits: nielsle_ (~nielsle@3239059-cl69.boa.fiberby.dk) (*.net *.split)</li>
 <li id="l-317"><a href="#l-317">#</a> [14:30] * Quits: Scorchin (u1242@gateway/web/irccloud.com/x-ykrsolpaiznltdvd) (*.net *.split)</li>
 <li id="l-318"><a href="#l-318">#</a> [14:30] * Quits: hamaji (~hamaji@220.109.219.244) (*.net *.split)</li>
 <li id="l-319"><a href="#l-319">#</a> [14:30] * Quits: othree (~othree@admin39.ct.ntust.edu.tw) (*.net *.split)</li>
 <li id="l-320"><a href="#l-320">#</a> [14:30] * Quits: gsnedders (~gsnedders@mail.gsnedders.com) (*.net *.split)</li>
 <li id="l-321"><a href="#l-321">#</a> [14:30] * Quits: eighty4 (~eighty4@unaffiliated/eighty4) (*.net *.split)</li>
 <li id="l-322"><a href="#l-322">#</a> [14:30] * Quits: jmb (~jmb@mail.parsifal.org.uk) (*.net *.split)</li>
 <li id="l-323"><a href="#l-323">#</a> [14:30] * Quits: volkmar (~volkmar@gentoo/developer/volkmar) (*.net *.split)</li>
 <li id="l-324" _a="jarib_"><a href="#l-324">#</a> [14:30] * jarib_ is now known as jarib</li>
 <li id="l-325" _a="tomaw"><a href="#l-325">#</a> [14:35] * tomaw is now known as 13WAATSDK</li>
 <li id="l-326" _a="jarib"><a href="#l-326">#</a> [14:35] * jarib is now known as 13WAAQZXE</li>
 <li id="l-327"><a href="#l-327">#</a> [14:36] * Joins: nlogax (~nlogax@unaffiliated/nlogax)</li>
 <li id="l-328"><a href="#l-328">#</a> [14:36] * Joins: Taggnostr (~quassel@dyn57-215.yok.fi)</li>
 <li id="l-329"><a href="#l-329">#</a> [14:36] * Joins: lhnz (~lhnz@188-223-83-48.zone14.bethere.co.uk)</li>
 <li id="l-330"><a href="#l-330">#</a> [14:36] * Joins: pererik (~pe@unaffiliated/pererik)</li>
 <li id="l-331"><a href="#l-331">#</a> [14:36] * Joins: bensmithett (~bensmithe@115.146.71.1)</li>
 <li id="l-332"><a href="#l-332">#</a> [14:36] * Joins: cgcardona (~cgcardona@unaffiliated/cgcardona)</li>
 <li id="l-333"><a href="#l-333">#</a> [14:36] * Joins: purefusion (~purefusio@unaffiliated/purefusion)</li>
 <li id="l-334"><a href="#l-334">#</a> [14:36] * Joins: llrcombs (~llrcombs@64.130.210.214)</li>
 <li id="l-335"><a href="#l-335">#</a> [14:36] * Joins: Kingdutch (~kingdutch@188.200.149.217)</li>
 <li id="l-336"><a href="#l-336">#</a> [14:36] * Joins: danheberden (~danheberd@li225-35.members.linode.com)</li>
 <li id="l-337"><a href="#l-337">#</a> [14:36] * Joins: nunnun (~nunnun@2001:268:355:1:20c:29ff:fed5:973a)</li>
 <li id="l-338"><a href="#l-338">#</a> [14:36] * Joins: silky (~silky@pool-74-108-142-22.nycmny.fios.verizon.net)</li>
 <li id="l-339"><a href="#l-339">#</a> [14:36] * Joins: hoodow (~hoodow@pdpc/supporter/active/hoodow)</li>
 <li id="l-340"><a href="#l-340">#</a> [14:36] * Joins: bentruyman (~bentruyma@li159-104.members.linode.com)</li>
 <li id="l-341"><a href="#l-341">#</a> [14:36] * Joins: gavin (~gavin@firefox/developer/gavin)</li>
 <li id="l-342"><a href="#l-342">#</a> [14:36] * Joins: mven (~mven__@169.241.49.57)</li>
 <li id="l-343"><a href="#l-343">#</a> [14:36] * Joins: GPH-Zeke (~GPHemsley@pdpc/supporter/student/GPHemsley)</li>
 <li id="l-344"><a href="#l-344">#</a> [14:36] * Joins: Moo_O (~quassel@herd37.twinapex.fi)</li>
 <li id="l-345"><a href="#l-345">#</a> [14:36] * Joins: hasather (~davidh@pat-tdc.opera.com)</li>
 <li id="l-346"><a href="#l-346">#</a> [14:36] * Joins: paul_irish (~paul_iris@ve.hsh6wjwx.vesrv.com)</li>
 <li id="l-347"><a href="#l-347">#</a> [14:36] * Joins: ivan`` (~ivan@li125-242.members.linode.com)</li>
 <li id="l-348"><a href="#l-348">#</a> [14:36] * Joins: jarib (~jarib@unaffiliated/jarib)</li>
 <li id="l-349"><a href="#l-349">#</a> [14:36] * Joins: ParadoX- (parad0x@hades.spexhost.com)</li>
 <li id="l-350"><a href="#l-350">#</a> [14:36] * Joins: ben_alman_ (~ben_alman@web126.webfaction.com)</li>
 <li id="l-351"><a href="#l-351">#</a> [14:36] * Joins: doublec (~doublec@unaffiliated/doublec)</li>
 <li id="l-352"><a href="#l-352">#</a> [14:36] * Joins: hendry (~hendry@85.119.83.83)</li>
 <li id="l-353"><a href="#l-353">#</a> [14:36] * Joins: Velmont (odinho@knuth.ping.uio.no)</li>
 <li id="l-354"><a href="#l-354">#</a> [14:36] * Joins: toyoshim_ (~toyoshim@yuri.twintail.org)</li>
 <li id="l-355"><a href="#l-355">#</a> [14:36] * Joins: kbrosnan (~kbrosnan@firefox/community/qa/kbrosnan)</li>
 <li id="l-356"><a href="#l-356">#</a> [14:36] * Joins: Hixie (~ianh@trivini.no)</li>
 <li id="l-357"><a href="#l-357">#</a> [14:36] * Joins: wirepair (fbi@random.supermario.org)</li>
 <li id="l-358"><a href="#l-358">#</a> [14:36] * Joins: heycam|away (~cam@wok.mcc.id.au)</li>
 <li id="l-359"><a href="#l-359">#</a> [14:36] * Joins: manu-db (~msporny@digitalbazaar.com)</li>
 <li id="l-360"><a href="#l-360">#</a> [14:36] * Joins: niftylettuce (u2733@gateway/web/irccloud.com/x-ayzqrjvdyvlvqbpy)</li>
 <li id="l-361"><a href="#l-361">#</a> [14:36] * Joins: beowulf (u116@pdpc/supporter/professional/beowulf)</li>
 <li id="l-362"><a href="#l-362">#</a> [14:36] * Joins: bobylito (u3929@gateway/web/irccloud.com/x-cknembtuoifrrjlj)</li>
 <li id="l-363"><a href="#l-363">#</a> [14:36] * Joins: annevk5 (u2483@gateway/web/irccloud.com/x-mtcimozrgjewgrki)</li>
 <li id="l-364"><a href="#l-364">#</a> [14:36] * Joins: eric_carlson (~ericc@adsl-67-112-12-110.dsl.anhm01.pacbell.net)</li>
 <li id="l-365"><a href="#l-365">#</a> [14:36] * Joins: franksalim (~frank@64-71-23-251.static.wiline.com)</li>
 <li id="l-366"><a href="#l-366">#</a> [14:36] * Joins: Yudai (~Yudai@p656629.tokynt01.ap.so-net.ne.jp)</li>
 <li id="l-367"><a href="#l-367">#</a> [14:36] * Joins: zewt (~x@c-24-62-196-44.hsd1.ma.comcast.net)</li>
 <li id="l-368"><a href="#l-368">#</a> [14:36] * Joins: jennb (~jennb@74.125.59.65)</li>
 <li id="l-369"><a href="#l-369">#</a> [14:36] * Joins: bzed (~bzed@devel.recluse.de)</li>
 <li id="l-370"><a href="#l-370">#</a> [14:36] * Joins: payman (~payman@pat.se.opera.com)</li>
 <li id="l-371"><a href="#l-371">#</a> [14:36] * Joins: jernoble|afk (~jernoble@17.212.152.13)</li>
 <li id="l-372"><a href="#l-372">#</a> [14:36] * Joins: Peter` (~peter@nishino.lvp-media.com)</li>
 <li id="l-373"><a href="#l-373">#</a> [14:36] * Joins: foolip_ (u3586@gateway/web/irccloud.com/x-cjhbzbmtwahekqba)</li>
 <li id="l-374"><a href="#l-374">#</a> [14:36] * Joins: _salavas (~salavas@c83-248-102-83.bredband.comhem.se)</li>
 <li id="l-375"><a href="#l-375">#</a> [14:36] * Joins: benjoffe_ (~benjoffe_@CPE-121-216-39-241.lnse1.ken.bigpond.net.au)</li>
 <li id="l-376"><a href="#l-376">#</a> [14:36] * Joins: astearns (~anonymous@c-50-132-63-33.hsd1.wa.comcast.net)</li>
 <li id="l-377"><a href="#l-377">#</a> [14:36] * Joins: danielfilho (~daniel@187.31.77.7)</li>
 <li id="l-378"><a href="#l-378">#</a> [14:36] * Joins: svl (~me@ip565744a7.direct-adsl.nl)</li>
 <li id="l-379"><a href="#l-379">#</a> [14:36] * Joins: saba (~foo@unaffiliated/saba)</li>
 <li id="l-380"><a href="#l-380">#</a> [14:36] * Joins: drublic (~drublic@frbg-5d84ed6c.pool.mediaWays.net)</li>
 <li id="l-381"><a href="#l-381">#</a> [14:36] * Joins: Telling (~unknown@80-71-135-15.u.parknet.dk)</li>
 <li id="l-382"><a href="#l-382">#</a> [14:36] * Joins: JuanDaugherty (~Ren@cpe-72-228-177-92.buffalo.res.rr.com)</li>
 <li id="l-383"><a href="#l-383">#</a> [14:36] * Joins: Maurice (copyman@5ED573FA.cm-7-6b.dynamic.ziggo.nl)</li>
 <li id="l-384"><a href="#l-384">#</a> [14:36] * Joins: dglazkov (u4270@gateway/web/irccloud.com/x-eteinghaavexdwwy)</li>
 <li id="l-385"><a href="#l-385">#</a> [14:36] * Joins: FireFly (firefly@unaffiliated/firefly)</li>
 <li id="l-386"><a href="#l-386">#</a> [14:36] * Joins: crash\ (bouncer@lubyte.de)</li>
 <li id="l-387"><a href="#l-387">#</a> [14:36] * Joins: Workshiva (~Dashiva@74.125.57.33)</li>
 <li id="l-388"><a href="#l-388">#</a> [14:36] * Joins: connrs (~connrs@conners.plus.com)</li>
 <li id="l-389"><a href="#l-389">#</a> [14:36] * Joins: stalled (~stalled@unaffiliated/stalled)</li>
 <li id="l-390"><a href="#l-390">#</a> [14:36] * Joins: shetech (~shetech@c-76-126-167-49.hsd1.ca.comcast.net)</li>
 <li id="l-391"><a href="#l-391">#</a> [14:36] * Joins: _bga (~bga@95-55-61-178.dynamic.avangarddsl.ru)</li>
 <li id="l-392"><a href="#l-392">#</a> [14:36] * Joins: krijn (u2319@gateway/web/irccloud.com/x-gdyudgtmfzwgpoei)</li>
 <li id="l-393"><a href="#l-393">#</a> [14:36] * Joins: lensco (u4054@gateway/web/irccloud.com/x-smwpdzgtvjkkbdni)</li>
 <li id="l-394"><a href="#l-394">#</a> [14:36] * Joins: antti_s (~asal@backport.reaktor.fi)</li>
 <li id="l-395"><a href="#l-395">#</a> [14:36] * Joins: hober (~ted@unaffiliated/hober)</li>
 <li id="l-396"><a href="#l-396">#</a> [14:36] * Joins: webben (~benjamin@173-203-84-17.static.cloud-ips.com)</li>
 <li id="l-397"><a href="#l-397">#</a> [14:36] * Joins: volkmar (~volkmar@gentoo/developer/volkmar)</li>
 <li id="l-398"><a href="#l-398">#</a> [14:36] * Joins: jmb (~jmb@mail.parsifal.org.uk)</li>
 <li id="l-399"><a href="#l-399">#</a> [14:36] * Joins: eighty4 (~eighty4@unaffiliated/eighty4)</li>
 <li id="l-400"><a href="#l-400">#</a> [14:36] * Joins: gsnedders (~gsnedders@mail.gsnedders.com)</li>
 <li id="l-401"><a href="#l-401">#</a> [14:36] * Joins: othree (~othree@admin39.ct.ntust.edu.tw)</li>
 <li id="l-402"><a href="#l-402">#</a> [14:36] * Joins: hamaji (~hamaji@220.109.219.244)</li>
 <li id="l-403"><a href="#l-403">#</a> [14:36] * Joins: Scorchin (u1242@gateway/web/irccloud.com/x-ykrsolpaiznltdvd)</li>
 <li id="l-404"><a href="#l-404">#</a> [14:36] * Joins: nielsle_ (~nielsle@3239059-cl69.boa.fiberby.dk)</li>
 <li id="l-405"><a href="#l-405">#</a> [14:36] * Joins: foolip (~philip@83.218.67.122)</li>
 <li id="l-406"><a href="#l-406">#</a> [14:36] * Joins: slartsa (~slartsa@alpha.pumppumedia.com)</li>
 <li id="l-407"><a href="#l-407">#</a> [14:36] * Joins: riven (~riven@pdpc/supporter/professional/riven)</li>
 <li id="l-408"><a href="#l-408">#</a> [14:36] * Joins: tmzt (~tmzt@adsl-99-164-105-221.dsl.akrnoh.sbcglobal.net)</li>
 <li id="l-409"><a href="#l-409">#</a> [14:36] * Joins: annevk (~annevk@5355737B.cm-6-6b.dynamic.ziggo.nl)</li>
 <li id="l-410"><a href="#l-410">#</a> [14:36] * Joins: jgraham (~jgraham@74.53.238.210)</li>
 <li id="l-411"><a href="#l-411">#</a> [14:36] * Joins: cpearce (~chatzilla@ip-118-90-36-154.xdsl.xnet.co.nz)</li>
 <li id="l-412"><a href="#l-412">#</a> [14:36] * Joins: PalleZingmark (~Adium@5e0571a8.bb.sky.com)</li>
 <li id="l-413"><a href="#l-413">#</a> [14:36] * Joins: reggna (~reggna@godis.olf.sgsnet.se)</li>
 <li id="l-414"><a href="#l-414">#</a> [14:36] * Joins: hsivonen (~hsivonen@srv-e205.esp.mediateam.fi)</li>
 <li id="l-415"><a href="#l-415">#</a> [14:36] * Joins: matjas (u2247@gateway/web/irccloud.com/x-wrsfupcozgamgysp)</li>
 <li id="l-416"><a href="#l-416">#</a> [14:36] * Joins: inimino (~inimino@69.164.198.129)</li>
 <li id="l-417"><a href="#l-417">#</a> [14:36] * Joins: TabAtkins (~tabatkins@216.239.45.4)</li>
 <li id="l-418"><a href="#l-418">#</a> [14:36] * Joins: Philip` (~philip@92.243.11.39)</li>
 <li id="l-419"><a href="#l-419">#</a> [14:36] * Joins: gavin_ (~gavin@76.14.70.183)</li>
 <li id="l-420"><a href="#l-420">#</a> [14:36] * Joins: Phae (u455@gateway/web/irccloud.com/x-kjkaumlrcwwsuwop)</li>
 <li id="l-421"><a href="#l-421">#</a> [14:36] * Joins: plutoniix (~plutoniix@ppp-110-169-248-81.revip5.asianet.co.th)</li>
 <li id="l-422"><a href="#l-422">#</a> [14:36] * Joins: broquaint (a52b469f90@78.47.79.137)</li>
 <li id="l-423"><a href="#l-423">#</a> [14:36] * Joins: tellnes (~KrooniX@ec2-79-125-26-36.eu-west-1.compute.amazonaws.com)</li>
 <li id="l-424"><a href="#l-424">#</a> [14:36] * Joins: Mike[tm] (~MikeSmith@sideshowbarker.net)</li>
 <li id="l-425"><a href="#l-425">#</a> [14:36] * Joins: Martijnc (~Martijn@d54C02C64.access.telenet.be)</li>
 <li id="l-426"><a href="#l-426">#</a> [14:36] * Joins: asmodai (asmodai@178-85-121-247.dynamic.upc.nl)</li>
 <li id="l-427"><a href="#l-427">#</a> [14:36] * Joins: esc__ (~esc-ape@193.170.99.241)</li>
 <li id="l-428"><a href="#l-428">#</a> [14:36] * Joins: FlorianX (~Dimitri@p4FCF7A7E.dip.t-dialin.net)</li>
 <li id="l-429"><a href="#l-429">#</a> [14:36] * Joins: ukai (ukai@nat/google/x-sufkgyystqmagdvi)</li>
 <li id="l-430"><a href="#l-430">#</a> [14:36] * Joins: jcranmer (~jcranmer@ltsp2.csl.tjhsst.edu)</li>
 <li id="l-431"><a href="#l-431">#</a> [14:36] * Joins: Jedi_ (~Jedi@jedi.org)</li>
 <li id="l-432"><a href="#l-432">#</a> [14:36] * Joins: matijsb (u2278@gateway/web/irccloud.com/x-abqdczhjtjictiww)</li>
 <li id="l-433"><a href="#l-433">#</a> [14:36] * Joins: manu` (~chatzilla@pool-74-107-162-72.ronkva.east.verizon.net)</li>
 <li id="l-434"><a href="#l-434">#</a> [14:36] * Joins: michel_v (~tofu@unaffiliated/michelv/x-000000001)</li>
 <li id="l-435"><a href="#l-435">#</a> [14:36] * Joins: kinetik (~kinetik@121.98.132.55)</li>
 <li id="l-436"><a href="#l-436">#</a> [14:36] * Joins: eae (u4278@gateway/web/irccloud.com/x-ptpufxwfrnbvkwnn)</li>
 <li id="l-437"><a href="#l-437">#</a> [14:36] * Joins: wycats (u79@gateway/web/irccloud.com/x-uhodpmtelvvwclku)</li>
 <li id="l-438"><a href="#l-438">#</a> [14:36] * Joins: jeremyselier (u2513@gateway/web/irccloud.com/x-mboguljwqhlljoym)</li>
 <li id="l-439"><a href="#l-439">#</a> [14:36] * Joins: romainhuet (u2533@gateway/web/irccloud.com/x-mdxlmkhggzrtgfrt)</li>
 <li id="l-440"><a href="#l-440">#</a> [14:36] * Joins: AryehGregor (~Simetrica@mediawiki/simetrical)</li>
 <li id="l-441"><a href="#l-441">#</a> [14:36] * Joins: kennyluck (~kennyluck@114-43-120-164.dynamic.hinet.net)</li>
 <li id="l-442"><a href="#l-442">#</a> [14:36] * Joins: Morphous (jan@unaffiliated/amorphous)</li>
 <li id="l-443"><a href="#l-443">#</a> [14:36] * Joins: temp01 (~temp01@unaffiliated/temp01)</li>
 <li id="l-444"><a href="#l-444">#</a> [14:36] * Joins: Ms2ger (~Ms2ger@91.181.168.54)</li>
 <li id="l-445"><a href="#l-445">#</a> [14:36] * Joins: jarek (~jarek@unaffiliated/jarek)</li>
 <li id="l-446"><a href="#l-446">#</a> [14:36] * Joins: erlehmann (~erlehmann@82.113.121.67)</li>
 <li id="l-447"><a href="#l-447">#</a> [14:36] * Joins: andyg (~andyg@CPE-124-189-148-81.sqcy1.win.bigpond.net.au)</li>
 <li id="l-448"><a href="#l-448">#</a> [14:36] * Joins: tndH (~Rob@cpc16-seac19-2-0-cust234.7-2.cable.virginmedia.com)</li>
 <li id="l-449"><a href="#l-449">#</a> [14:36] * Joins: tomasf (~tom@c-b7dbe555.024-204-6c6b7012.cust.bredbandsbolaget.se)</li>
 <li id="l-450"><a href="#l-450">#</a> [14:36] * Joins: MikeSmith (~MikeSmith@EM1-113-48-31.pool.e-mobile.ne.jp)</li>
 <li id="l-451"><a href="#l-451">#</a> [14:36] * Joins: nonge__ (~nonge@p5B326AF0.dip.t-dialin.net)</li>
 <li id="l-452"><a href="#l-452">#</a> [14:36] * Joins: jacobolus (~jacobolus@c-71-198-169-213.hsd1.ca.comcast.net)</li>
 <li id="l-453"><a href="#l-453">#</a> [14:36] * Joins: gkellogg (~gregg@c-98-248-150-91.hsd1.ca.comcast.net)</li>
 <li id="l-454"><a href="#l-454">#</a> [14:36] * Joins: Necrathex (~nectop@82-170-160-25.ip.telfort.nl)</li>
 <li id="l-455"><a href="#l-455">#</a> [14:36] * Joins: arv (u4269@gateway/web/irccloud.com/x-hjpsmrvjscknpads)</li>
 <li id="l-456"><a href="#l-456">#</a> [14:36] * Joins: yutak (~yutak@2401:fa00:4:1004:baac:6fff:fe99:adfb)</li>
 <li id="l-457"><a href="#l-457">#</a> [14:36] * Joins: PrgmrBill (~PrgmrBill@unaffiliated/prgmrbill)</li>
 <li id="l-458"><a href="#l-458">#</a> [14:36] * Joins: gwillen (~gwillen@unaffiliated/gwillen)</li>
 <li id="l-459"><a href="#l-459">#</a> [14:36] * Joins: Rubennn (~quassel@2a02:348:33:5823::1)</li>
 <li id="l-460"><a href="#l-460">#</a> [14:36] * Joins: rektide (~rektide@67.217.170.145)</li>
 <li id="l-461"><a href="#l-461">#</a> [14:36] * Joins: boblet (u1921@gateway/web/irccloud.com/x-rckvknshduohrcqr)</li>
 <li id="l-462"><a href="#l-462">#</a> [14:36] * Joins: Obvious_MkII (tachikoma@188.226.74.2)</li>
 <li id="l-463"><a href="#l-463">#</a> [14:36] * Joins: Russ_ (~quassel@nat/transgaming/x-llrapuqqiqvpahbf)</li>
 <li id="l-464"><a href="#l-464">#</a> [14:36] * Joins: michaelw (~michaelw@2a01:4f8:131:23c1::5)</li>
 <li id="l-465"><a href="#l-465">#</a> [14:36] * Joins: wilhelm (~wilhelm@178.255.149.98)</li>
 <li id="l-466"><a href="#l-466">#</a> [14:36] * Joins: nw` (eero@heaven.unlink.org)</li>
 <li id="l-467"><a href="#l-467">#</a> [14:36] * Joins: Dashiva (Dashiva@wikia/Dashiva)</li>
 <li id="l-468"><a href="#l-468">#</a> [14:36] * Joins: MrOpposite (~MrOpposit@unaffiliated/mropposite)</li>
 <li id="l-469"><a href="#l-469">#</a> [14:36] * Joins: djr (~djr@208-75-85-139.slicehost.net)</li>
 <li id="l-470"><a href="#l-470">#</a> [14:36] * Joins: Druid_ (~Druid@p5B1376A3.dip.t-dialin.net)</li>
 <li id="l-471" _a="13WAATSDK"><a href="#l-471">#</a> [14:41] * 13WAATSDK is now known as tomaw</li>
 <li id="l-472"><a href="#l-472">#</a> [14:49] * Quits: Martijnc (~Martijn@d54C02C64.access.telenet.be) (Ping timeout: 260 seconds)</li>
 <li id="l-473"><a href="#l-473">#</a> [14:49] * Quits: Maurice (copyman@5ED573FA.cm-7-6b.dynamic.ziggo.nl) (*.net *.split)</li>
 <li id="l-474"><a href="#l-474">#</a> [14:49] * Quits: dglazkov (u4270@gateway/web/irccloud.com/x-eteinghaavexdwwy) (*.net *.split)</li>
 <li id="l-475"><a href="#l-475">#</a> [14:49] * Quits: FireFly (firefly@unaffiliated/firefly) (*.net *.split)</li>
 <li id="l-476"><a href="#l-476">#</a> [14:49] * Quits: crash\ (bouncer@lubyte.de) (*.net *.split)</li>
 <li id="l-477"><a href="#l-477">#</a> [14:49] * Quits: Workshiva (~Dashiva@74.125.57.33) (*.net *.split)</li>
 <li id="l-478"><a href="#l-478">#</a> [14:49] * Quits: cgcardona (~cgcardona@unaffiliated/cgcardona) (*.net *.split)</li>
 <li id="l-479"><a href="#l-479">#</a> [14:49] * Quits: bensmithett (~bensmithe@115.146.71.1) (*.net *.split)</li>
 <li id="l-480"><a href="#l-480">#</a> [14:49] * Quits: pererik (~pe@unaffiliated/pererik) (*.net *.split)</li>
 <li id="l-481"><a href="#l-481">#</a> [14:49] * Quits: lhnz (~lhnz@188-223-83-48.zone14.bethere.co.uk) (*.net *.split)</li>
 <li id="l-482"><a href="#l-482">#</a> [14:49] * Quits: Taggnostr (~quassel@dyn57-215.yok.fi) (*.net *.split)</li>
 <li id="l-483"><a href="#l-483">#</a> [14:49] * Quits: nlogax (~nlogax@unaffiliated/nlogax) (*.net *.split)</li>
 <li id="l-484"><a href="#l-484">#</a> [14:49] * Quits: JuanDaugherty (~Ren@cpe-72-228-177-92.buffalo.res.rr.com) (*.net *.split)</li>
 <li id="l-485"><a href="#l-485">#</a> [14:49] * Quits: svl (~me@ip565744a7.direct-adsl.nl) (*.net *.split)</li>
 <li id="l-486"><a href="#l-486">#</a> [14:49] * Quits: astearns (~anonymous@c-50-132-63-33.hsd1.wa.comcast.net) (*.net *.split)</li>
 <li id="l-487"><a href="#l-487">#</a> [14:49] * Quits: bzed (~bzed@devel.recluse.de) (*.net *.split)</li>
 <li id="l-488"><a href="#l-488">#</a> [14:49] * Quits: bobylito (u3929@gateway/web/irccloud.com/x-cknembtuoifrrjlj) (*.net *.split)</li>
 <li id="l-489"><a href="#l-489">#</a> [14:49] * Quits: heycam|away (~cam@wok.mcc.id.au) (*.net *.split)</li>
 <li id="l-490"><a href="#l-490">#</a> [14:49] * Quits: wirepair (fbi@random.supermario.org) (*.net *.split)</li>
 <li id="l-491"><a href="#l-491">#</a> [14:49] * Quits: ivan`` (~ivan@li125-242.members.linode.com) (*.net *.split)</li>
 <li id="l-492"><a href="#l-492">#</a> [14:49] * Quits: hasather (~davidh@pat-tdc.opera.com) (*.net *.split)</li>
 <li id="l-493"><a href="#l-493">#</a> [14:49] * Quits: mven (~mven__@169.241.49.57) (*.net *.split)</li>
 <li id="l-494"><a href="#l-494">#</a> [14:49] * Quits: llrcombs (~llrcombs@64.130.210.214) (*.net *.split)</li>
 <li id="l-495"><a href="#l-495">#</a> [14:49] * Quits: purefusion (~purefusio@unaffiliated/purefusion) (*.net *.split)</li>
 <li id="l-496"><a href="#l-496">#</a> [14:49] * Quits: danielfilho (~daniel@187.31.77.7) (*.net *.split)</li>
 <li id="l-497"><a href="#l-497">#</a> [14:49] * Quits: _salavas (~salavas@c83-248-102-83.bredband.comhem.se) (*.net *.split)</li>
 <li id="l-498"><a href="#l-498">#</a> [14:49] * Quits: payman (~payman@pat.se.opera.com) (*.net *.split)</li>
 <li id="l-499"><a href="#l-499">#</a> [14:49] * Quits: jennb (~jennb@74.125.59.65) (*.net *.split)</li>
 <li id="l-500"><a href="#l-500">#</a> [14:49] * Quits: zewt (~x@c-24-62-196-44.hsd1.ma.comcast.net) (*.net *.split)</li>
 <li id="l-501"><a href="#l-501">#</a> [14:49] * Quits: franksalim (~frank@64-71-23-251.static.wiline.com) (*.net *.split)</li>
 <li id="l-502"><a href="#l-502">#</a> [14:49] * Quits: eric_carlson (~ericc@adsl-67-112-12-110.dsl.anhm01.pacbell.net) (*.net *.split)</li>
 <li id="l-503"><a href="#l-503">#</a> [14:49] * Quits: annevk5 (u2483@gateway/web/irccloud.com/x-mtcimozrgjewgrki) (*.net *.split)</li>
 <li id="l-504"><a href="#l-504">#</a> [14:49] * Quits: manu-db (~msporny@digitalbazaar.com) (*.net *.split)</li>
 <li id="l-505"><a href="#l-505">#</a> [14:49] * Quits: Hixie (~ianh@trivini.no) (*.net *.split)</li>
 <li id="l-506"><a href="#l-506">#</a> [14:49] * Quits: toyoshim_ (~toyoshim@yuri.twintail.org) (*.net *.split)</li>
 <li id="l-507"><a href="#l-507">#</a> [14:49] * Quits: paul_irish (~paul_iris@ve.hsh6wjwx.vesrv.com) (*.net *.split)</li>
 <li id="l-508"><a href="#l-508">#</a> [14:49] * Quits: GPH-Zeke (~GPHemsley@pdpc/supporter/student/GPHemsley) (*.net *.split)</li>
 <li id="l-509"><a href="#l-509">#</a> [14:49] * Quits: Moo_O (~quassel@herd37.twinapex.fi) (*.net *.split)</li>
 <li id="l-510"><a href="#l-510">#</a> [14:49] * Quits: Telling (~unknown@80-71-135-15.u.parknet.dk) (*.net *.split)</li>
 <li id="l-511"><a href="#l-511">#</a> [14:49] * Quits: saba (~foo@unaffiliated/saba) (*.net *.split)</li>
 <li id="l-512"><a href="#l-512">#</a> [14:49] * Quits: foolip_ (u3586@gateway/web/irccloud.com/x-cjhbzbmtwahekqba) (*.net *.split)</li>
 <li id="l-513"><a href="#l-513">#</a> [14:49] * Quits: kbrosnan (~kbrosnan@firefox/community/qa/kbrosnan) (*.net *.split)</li>
 <li id="l-514"><a href="#l-514">#</a> [14:49] * Quits: doublec (~doublec@unaffiliated/doublec) (*.net *.split)</li>
 <li id="l-515"><a href="#l-515">#</a> [14:49] * Quits: jarib (~jarib@unaffiliated/jarib) (*.net *.split)</li>
 <li id="l-516"><a href="#l-516">#</a> [14:49] * Quits: hoodow (~hoodow@pdpc/supporter/active/hoodow) (*.net *.split)</li>
 <li id="l-517"><a href="#l-517">#</a> [14:49] * Quits: nunnun (~nunnun@2001:268:355:1:20c:29ff:fed5:973a) (*.net *.split)</li>
 <li id="l-518"><a href="#l-518">#</a> [14:49] * Quits: drublic (~drublic@frbg-5d84ed6c.pool.mediaWays.net) (*.net *.split)</li>
 <li id="l-519"><a href="#l-519">#</a> [14:49] * Quits: benjoffe_ (~benjoffe_@CPE-121-216-39-241.lnse1.ken.bigpond.net.au) (*.net *.split)</li>
 <li id="l-520"><a href="#l-520">#</a> [14:49] * Quits: Peter` (~peter@nishino.lvp-media.com) (*.net *.split)</li>
 <li id="l-521"><a href="#l-521">#</a> [14:49] * Quits: jernoble|afk (~jernoble@17.212.152.13) (*.net *.split)</li>
 <li id="l-522"><a href="#l-522">#</a> [14:49] * Quits: Yudai (~Yudai@p656629.tokynt01.ap.so-net.ne.jp) (*.net *.split)</li>
 <li id="l-523"><a href="#l-523">#</a> [14:49] * Quits: beowulf (u116@pdpc/supporter/professional/beowulf) (*.net *.split)</li>
 <li id="l-524"><a href="#l-524">#</a> [14:49] * Quits: niftylettuce (u2733@gateway/web/irccloud.com/x-ayzqrjvdyvlvqbpy) (*.net *.split)</li>
 <li id="l-525"><a href="#l-525">#</a> [14:49] * Quits: Velmont (odinho@knuth.ping.uio.no) (*.net *.split)</li>
 <li id="l-526"><a href="#l-526">#</a> [14:49] * Quits: hendry (~hendry@85.119.83.83) (*.net *.split)</li>
 <li id="l-527"><a href="#l-527">#</a> [14:49] * Quits: ben_alman_ (~ben_alman@web126.webfaction.com) (*.net *.split)</li>
 <li id="l-528"><a href="#l-528">#</a> [14:49] * Quits: ParadoX- (parad0x@hades.spexhost.com) (*.net *.split)</li>
 <li id="l-529"><a href="#l-529">#</a> [14:49] * Quits: gavin (~gavin@firefox/developer/gavin) (*.net *.split)</li>
 <li id="l-530"><a href="#l-530">#</a> [14:49] * Quits: bentruyman (~bentruyma@li159-104.members.linode.com) (*.net *.split)</li>
 <li id="l-531"><a href="#l-531">#</a> [14:49] * Quits: silky (~silky@pool-74-108-142-22.nycmny.fios.verizon.net) (*.net *.split)</li>
 <li id="l-532"><a href="#l-532">#</a> [14:49] * Quits: danheberden (~danheberd@li225-35.members.linode.com) (*.net *.split)</li>
 <li id="l-533"><a href="#l-533">#</a> [14:49] * Quits: Kingdutch (~kingdutch@188.200.149.217) (*.net *.split)</li>
 <li id="l-534"><a href="#l-534">#</a> [14:49] * Quits: connrs (~connrs@conners.plus.com) (*.net *.split)</li>
 <li id="l-535"><a href="#l-535">#</a> [14:49] * Quits: shetech (~shetech@c-76-126-167-49.hsd1.ca.comcast.net) (*.net *.split)</li>
 <li id="l-536"><a href="#l-536">#</a> [14:49] * Quits: _bga (~bga@95-55-61-178.dynamic.avangarddsl.ru) (*.net *.split)</li>
 <li id="l-537"><a href="#l-537">#</a> [14:49] * Quits: krijn (u2319@gateway/web/irccloud.com/x-gdyudgtmfzwgpoei) (*.net *.split)</li>
 <li id="l-538"><a href="#l-538">#</a> [14:49] * Quits: lensco (u4054@gateway/web/irccloud.com/x-smwpdzgtvjkkbdni) (*.net *.split)</li>
 <li id="l-539"><a href="#l-539">#</a> [14:49] * Quits: antti_s (~asal@backport.reaktor.fi) (*.net *.split)</li>
 <li id="l-540"><a href="#l-540">#</a> [14:49] * Quits: stalled (~stalled@unaffiliated/stalled) (*.net *.split)</li>
 <li id="l-541"><a href="#l-541">#</a> [14:49] * Quits: hober (~ted@unaffiliated/hober) (*.net *.split)</li>
 <li id="l-542"><a href="#l-542">#</a> [14:49] * Quits: webben (~benjamin@173-203-84-17.static.cloud-ips.com) (*.net *.split)</li>
 <li id="l-543"><a href="#l-543">#</a> [14:49] * Quits: Druid_ (~Druid@p5B1376A3.dip.t-dialin.net) (*.net *.split)</li>
 <li id="l-544"><a href="#l-544">#</a> [14:49] * Quits: FlorianX (~Dimitri@p4FCF7A7E.dip.t-dialin.net) (*.net *.split)</li>
 <li id="l-545"><a href="#l-545">#</a> [14:49] * Quits: esc__ (~esc-ape@193.170.99.241) (*.net *.split)</li>
 <li id="l-546"><a href="#l-546">#</a> [14:49] * Quits: asmodai (asmodai@178-85-121-247.dynamic.upc.nl) (*.net *.split)</li>
 <li id="l-547"><a href="#l-547">#</a> [14:49] * Quits: Mike[tm] (~MikeSmith@sideshowbarker.net) (*.net *.split)</li>
 <li id="l-548"><a href="#l-548">#</a> [14:49] * Quits: tellnes (~KrooniX@ec2-79-125-26-36.eu-west-1.compute.amazonaws.com) (*.net *.split)</li>
 <li id="l-549"><a href="#l-549">#</a> [14:49] * Quits: broquaint (a52b469f90@78.47.79.137) (*.net *.split)</li>
 <li id="l-550"><a href="#l-550">#</a> [14:50] * Joins: esc_ (~esc-ape@193.170.99.241)</li>
 <li id="l-551"><a href="#l-551">#</a> [14:50] * Joins: asmodai (asmodai@178-85-121-247.dynamic.upc.nl)</li>
 <li id="l-552"><a href="#l-552">#</a> [14:50] * Joins: FlorianX (~Dimitri@p4FCF7A7E.dip.t-dialin.net)</li>
 <li id="l-553" _a="tomaw"><a href="#l-553">#</a> [14:52] * tomaw is now known as 13WAATSDK</li>
 <li id="l-554"><a href="#l-554">#</a> [14:52] * Joins: KrooniX (~KrooniX@ec2-79-125-26-36.eu-west-1.compute.amazonaws.com)</li>
 <li id="l-555"><a href="#l-555">#</a> [14:52] * Joins: hober2 (~ted@unaffiliated/hober)</li>
 <li id="l-556"><a href="#l-556">#</a> [14:52] * Joins: broquaint (ad306d4d16@78.47.79.137)</li>
 <li id="l-557"><a href="#l-557">#</a> [14:52] * Joins: antti_s (~asal@backport.reaktor.fi)</li>
 <li id="l-558"><a href="#l-558">#</a> [14:52] * Joins: lensco (u4054@gateway/web/irccloud.com/x-smwpdzgtvjkkbdni)</li>
 <li id="l-559"><a href="#l-559">#</a> [14:52] * Joins: krijn (u2319@gateway/web/irccloud.com/x-gdyudgtmfzwgpoei)</li>
 <li id="l-560"><a href="#l-560">#</a> [14:52] * Joins: _bga (~bga@95-55-61-178.dynamic.avangarddsl.ru)</li>
 <li id="l-561"><a href="#l-561">#</a> [14:52] * Joins: shetech (~shetech@c-76-126-167-49.hsd1.ca.comcast.net)</li>
 <li id="l-562"><a href="#l-562">#</a> [14:52] * Joins: connrs (~connrs@conners.plus.com)</li>
 <li id="l-563"><a href="#l-563">#</a> [14:52] * Joins: JuanDaugherty (~Ren@cpe-72-228-177-92.buffalo.res.rr.com)</li>
 <li id="l-564"><a href="#l-564">#</a> [14:52] * Joins: Telling (~unknown@80-71-135-15.u.parknet.dk)</li>
 <li id="l-565"><a href="#l-565">#</a> [14:52] * Joins: drublic (~drublic@frbg-5d84ed6c.pool.mediaWays.net)</li>
 <li id="l-566"><a href="#l-566">#</a> [14:52] * Joins: saba (~foo@unaffiliated/saba)</li>
 <li id="l-567"><a href="#l-567">#</a> [14:52] * Joins: svl (~me@ip565744a7.direct-adsl.nl)</li>
 <li id="l-568"><a href="#l-568">#</a> [14:52] * Joins: danielfilho (~daniel@187.31.77.7)</li>
 <li id="l-569"><a href="#l-569">#</a> [14:52] * Joins: astearns (~anonymous@c-50-132-63-33.hsd1.wa.comcast.net)</li>
 <li id="l-570"><a href="#l-570">#</a> [14:52] * Joins: benjoffe_ (~benjoffe_@CPE-121-216-39-241.lnse1.ken.bigpond.net.au)</li>
 <li id="l-571"><a href="#l-571">#</a> [14:52] * Joins: _salavas (~salavas@c83-248-102-83.bredband.comhem.se)</li>
 <li id="l-572"><a href="#l-572">#</a> [14:52] * Joins: foolip_ (u3586@gateway/web/irccloud.com/x-cjhbzbmtwahekqba)</li>
 <li id="l-573"><a href="#l-573">#</a> [14:52] * Joins: Peter` (~peter@nishino.lvp-media.com)</li>
 <li id="l-574"><a href="#l-574">#</a> [14:52] * Joins: jernoble|afk (~jernoble@17.212.152.13)</li>
 <li id="l-575"><a href="#l-575">#</a> [14:52] * Joins: payman (~payman@pat.se.opera.com)</li>
 <li id="l-576"><a href="#l-576">#</a> [14:52] * Joins: bzed (~bzed@devel.recluse.de)</li>
 <li id="l-577"><a href="#l-577">#</a> [14:52] * Joins: jennb (~jennb@74.125.59.65)</li>
 <li id="l-578"><a href="#l-578">#</a> [14:52] * Joins: zewt (~x@c-24-62-196-44.hsd1.ma.comcast.net)</li>
 <li id="l-579"><a href="#l-579">#</a> [14:52] * Joins: Yudai (~Yudai@p656629.tokynt01.ap.so-net.ne.jp)</li>
 <li id="l-580"><a href="#l-580">#</a> [14:52] * Joins: franksalim (~frank@64-71-23-251.static.wiline.com)</li>
 <li id="l-581"><a href="#l-581">#</a> [14:52] * Joins: eric_carlson (~ericc@adsl-67-112-12-110.dsl.anhm01.pacbell.net)</li>
 <li id="l-582"><a href="#l-582">#</a> [14:52] * Joins: annevk5 (u2483@gateway/web/irccloud.com/x-mtcimozrgjewgrki)</li>
 <li id="l-583"><a href="#l-583">#</a> [14:52] * Joins: bobylito (u3929@gateway/web/irccloud.com/x-cknembtuoifrrjlj)</li>
 <li id="l-584"><a href="#l-584">#</a> [14:52] * Joins: beowulf (u116@pdpc/supporter/professional/beowulf)</li>
 <li id="l-585"><a href="#l-585">#</a> [14:52] * Joins: niftylettuce (u2733@gateway/web/irccloud.com/x-ayzqrjvdyvlvqbpy)</li>
 <li id="l-586"><a href="#l-586">#</a> [14:52] * Joins: manu-db (~msporny@digitalbazaar.com)</li>
 <li id="l-587"><a href="#l-587">#</a> [14:52] * Joins: heycam|away (~cam@wok.mcc.id.au)</li>
 <li id="l-588"><a href="#l-588">#</a> [14:52] * Joins: wirepair (fbi@random.supermario.org)</li>
 <li id="l-589"><a href="#l-589">#</a> [14:52] * Joins: Hixie (~ianh@trivini.no)</li>
 <li id="l-590"><a href="#l-590">#</a> [14:52] * Joins: kbrosnan (~kbrosnan@firefox/community/qa/kbrosnan)</li>
 <li id="l-591"><a href="#l-591">#</a> [14:52] * Joins: toyoshim_ (~toyoshim@yuri.twintail.org)</li>
 <li id="l-592"><a href="#l-592">#</a> [14:52] * Joins: Velmont (odinho@knuth.ping.uio.no)</li>
 <li id="l-593"><a href="#l-593">#</a> [14:52] * Joins: hendry (~hendry@85.119.83.83)</li>
 <li id="l-594"><a href="#l-594">#</a> [14:52] * Joins: doublec (~doublec@unaffiliated/doublec)</li>
 <li id="l-595"><a href="#l-595">#</a> [14:52] * Joins: ben_alman_ (~ben_alman@web126.webfaction.com)</li>
 <li id="l-596"><a href="#l-596">#</a> [14:52] * Joins: ParadoX- (parad0x@hades.spexhost.com)</li>
 <li id="l-597"><a href="#l-597">#</a> [14:52] * Joins: jarib (~jarib@unaffiliated/jarib)</li>
 <li id="l-598"><a href="#l-598">#</a> [14:52] * Joins: ivan`` (~ivan@li125-242.members.linode.com)</li>
 <li id="l-599"><a href="#l-599">#</a> [14:52] * Joins: paul_irish (~paul_iris@ve.hsh6wjwx.vesrv.com)</li>
 <li id="l-600"><a href="#l-600">#</a> [14:52] * Joins: hasather (~davidh@pat-tdc.opera.com)</li>
 <li id="l-601"><a href="#l-601">#</a> [14:52] * Joins: Moo_O (~quassel@herd37.twinapex.fi)</li>
 <li id="l-602"><a href="#l-602">#</a> [14:52] * Joins: GPH-Zeke (~GPHemsley@pdpc/supporter/student/GPHemsley)</li>
 <li id="l-603"><a href="#l-603">#</a> [14:52] * Joins: mven (~mven__@169.241.49.57)</li>
 <li id="l-604"><a href="#l-604">#</a> [14:52] * Joins: gavin (~gavin@firefox/developer/gavin)</li>
 <li id="l-605"><a href="#l-605">#</a> [14:52] * Joins: bentruyman (~bentruyma@li159-104.members.linode.com)</li>
 <li id="l-606"><a href="#l-606">#</a> [14:52] * Joins: hoodow (~hoodow@pdpc/supporter/active/hoodow)</li>
 <li id="l-607"><a href="#l-607">#</a> [14:52] * Joins: silky (~silky@pool-74-108-142-22.nycmny.fios.verizon.net)</li>
 <li id="l-608"><a href="#l-608">#</a> [14:52] * Joins: nunnun (~nunnun@2001:268:355:1:20c:29ff:fed5:973a)</li>
 <li id="l-609"><a href="#l-609">#</a> [14:52] * Joins: danheberden (~danheberd@li225-35.members.linode.com)</li>
 <li id="l-610"><a href="#l-610">#</a> [14:52] * Joins: Kingdutch (~kingdutch@188.200.149.217)</li>
 <li id="l-611"><a href="#l-611">#</a> [14:52] * Joins: llrcombs (~llrcombs@64.130.210.214)</li>
 <li id="l-612"><a href="#l-612">#</a> [14:52] * Joins: purefusion (~purefusio@unaffiliated/purefusion)</li>
 <li id="l-613"><a href="#l-613">#</a> [14:52] * Joins: cgcardona (~cgcardona@unaffiliated/cgcardona)</li>
 <li id="l-614"><a href="#l-614">#</a> [14:52] * Joins: bensmithett (~bensmithe@115.146.71.1)</li>
 <li id="l-615"><a href="#l-615">#</a> [14:52] * Joins: pererik (~pe@unaffiliated/pererik)</li>
 <li id="l-616"><a href="#l-616">#</a> [14:52] * Joins: lhnz (~lhnz@188-223-83-48.zone14.bethere.co.uk)</li>
 <li id="l-617"><a href="#l-617">#</a> [14:52] * Joins: Taggnostr (~quassel@dyn57-215.yok.fi)</li>
 <li id="l-618"><a href="#l-618">#</a> [14:52] * Joins: nlogax (~nlogax@unaffiliated/nlogax)</li>
 <li id="l-619"><a href="#l-619">#</a> [14:52] * Joins: Workshiva (~Dashiva@74.125.57.33)</li>
 <li id="l-620"><a href="#l-620">#</a> [14:52] * Joins: crash\ (bouncer@lubyte.de)</li>
 <li id="l-621"><a href="#l-621">#</a> [14:52] * Joins: FireFly (firefly@unaffiliated/firefly)</li>
 <li id="l-622"><a href="#l-622">#</a> [14:52] * Joins: dglazkov (u4270@gateway/web/irccloud.com/x-eteinghaavexdwwy)</li>
 <li id="l-623"><a href="#l-623">#</a> [14:52] * Joins: Maurice (copyman@5ED573FA.cm-7-6b.dynamic.ziggo.nl)</li>
 <li id="l-624"><a href="#l-624">#</a> [14:52] * Joins: [tm] (~MikeSmith@sideshowbarker.net)</li>
 <li id="l-625" _a="13WAATSDK"><a href="#l-625">#</a> [14:54] * 13WAATSDK is now known as tomaw</li>
 <li id="l-626"><a href="#l-626">#</a> [14:56] * Joins: stalled (~stalled@unaffiliated/stalled)</li>
 <li id="l-627"><a href="#l-627">#</a> [14:59] * Joins: Druid_ (~Druid@p5B1376A3.dip.t-dialin.net)</li>
 <li id="l-628"><a href="#l-628">#</a> [15:20] * Joins: webben (~benjamin@173-203-84-17.static.cloud-ips.com)</li>
 <li id="l-629"><a href="#l-629">#</a> [15:24] * Quits: jarek (~jarek@unaffiliated/jarek) (Quit: Leaving)</li>
 <li id="l-630"><a href="#l-630">#</a> [15:28] * Quits: PalleZingmark (~Adium@5e0571a8.bb.sky.com) (Quit: Leaving.)</li>
 <li id="l-631"><a href="#l-631">#</a> [16:44] * Joins: smaug____ (~chatzilla@GZKMMCMLVIII.gprs.sl-laajakaista.fi)</li>
 <li id="l-632"><a href="#l-632">#</a> [17:00] * Joins: Martijnc (~Martijn@d54C02C64.access.telenet.be)</li>
 <li id="l-633"><a href="#l-633">#</a> [17:03] * Quits: MikeSmith (~MikeSmith@EM1-113-48-31.pool.e-mobile.ne.jp) (Ping timeout: 252 seconds)</li>
 <li id="l-634"><a href="#l-634">#</a> [17:09] * Joins: MikeSmith (~MikeSmith@EM1-113-19-159.pool.e-mobile.ne.jp)</li>
 <li id="l-635"><a href="#l-635">#</a> [17:23] * Quits: temp01 (~temp01@unaffiliated/temp01) (Ping timeout: 240 seconds)</li>
 <li id="l-636"><a href="#l-636">#</a> [17:24] * Quits: smaug____ (~chatzilla@GZKMMCMLVIII.gprs.sl-laajakaista.fi) (Ping timeout: 260 seconds)</li>
 <li id="l-637"><a href="#l-637">#</a> [17:25] * Joins: temp01 (~temp01@unaffiliated/temp01)</li>
 <li id="l-638"><a href="#l-638">#</a> [17:31] * Joins: tantek (~tantek@70-36-139-219.dsl.dynamic.sonic.net)</li>
 <li id="l-639"><a href="#l-639">#</a> [17:53] * Quits: jeremyselier (u2513@gateway/web/irccloud.com/x-mboguljwqhlljoym) (Quit: Connection closed for inactivity)</li>
 <li id="l-640"><a href="#l-640">#</a> [17:54] * Joins: bastones (~bastones@host86-145-136-236.range86-145.btcentralplus.com)</li>
 <li id="l-641"><a href="#l-641">#</a> [17:54] * Joins: smaug____ (~chatzilla@cs181151161.pp.htv.fi)</li>
 <li id="l-642"><a href="#l-642">#</a> [18:15] * Quits: bastones (~bastones@host86-145-136-236.range86-145.btcentralplus.com) (Remote host closed the connection)</li>
 <li id="l-643"><a href="#l-643">#</a> [18:16] * Quits: MikeSmith (~MikeSmith@EM1-113-19-159.pool.e-mobile.ne.jp) (Ping timeout: 248 seconds)</li>
 <li id="l-644"><a href="#l-644">#</a> [18:30] * Joins: jarek (~jarek@aeal107.neoplus.adsl.tpnet.pl)</li>
 <li id="l-645"><a href="#l-645">#</a> [18:30] * Quits: jarek (~jarek@aeal107.neoplus.adsl.tpnet.pl) (Changing host)</li>
 <li id="l-646"><a href="#l-646">#</a> [18:30] * Joins: jarek (~jarek@unaffiliated/jarek)</li>
 <li id="l-647"><a href="#l-647">#</a> [18:35] * Quits: andyg (~andyg@CPE-124-189-148-81.sqcy1.win.bigpond.net.au) (Quit: andyg)</li>
 <li id="l-648"><a href="#l-648">#</a> [18:38] * Quits: tndH (~Rob@cpc16-seac19-2-0-cust234.7-2.cable.virginmedia.com) (Remote host closed the connection)</li>
 <li id="l-649"><a href="#l-649">#</a> [18:51] * Quits: tantek (~tantek@70-36-139-219.dsl.dynamic.sonic.net) (Quit: tantek)</li>
 <li id="l-650"><a href="#l-650">#</a> [18:53] * Quits: jarek (~jarek@unaffiliated/jarek) (Ping timeout: 240 seconds)</li>
 <li id="l-651" _a="AryehGregor"><a href="#l-651">#</a> [18:54] &lt;AryehGregor&gt; Is there a good way to test [[Configurable]]?</li>
 <li id="l-652" _a="AryehGregor"><a href="#l-652">#</a> [18:54] &lt;AryehGregor&gt; Just delete and see if it works?</li>
 <li id="l-653" _a="_bga"><a href="#l-653">#</a> [18:54] &lt;_bga&gt; getOPD?</li>
 <li id="l-654" _a="AryehGregor" _r="Oh"><a href="#l-654">#</a> [18:55] &lt;AryehGregor&gt; Oh, awesome.</li>
 <li id="l-655" _a="AryehGregor"><a href="#l-655">#</a> [18:55] &lt;AryehGregor&gt; Thanks!</li>
 <li id="l-656" _a="AryehGregor"><a href="#l-656">#</a> [18:56] &lt;AryehGregor&gt; I guess if that returns the right thing but it doesn't behave right, that's more of an ES bug than a WebIDL bug.</li>
 <li id="l-657"><a href="#l-657">#</a> [18:59] * Joins: mpt (~mpt@76.14.69.234)</li>
 <li id="l-658"><a href="#l-658">#</a> [18:59] * Quits: mpt (~mpt@76.14.69.234) (Changing host)</li>
 <li id="l-659"><a href="#l-659">#</a> [18:59] * Joins: mpt (~mpt@canonical/mpt)</li>
 <li id="l-660"><a href="#l-660">#</a> [19:05] * Quits: plutoniix (~plutoniix@ppp-110-169-248-81.revip5.asianet.co.th) (Quit: Leaving)</li>
 <li id="l-661" _a="_bga"><a href="#l-661">#</a> [19:09] &lt;_bga&gt; AryehGregor <a href="http://pastie.org/2857719">http://pastie.org/2857719</a></li>
 <li id="l-662" _a="AryehGregor"><a href="#l-662">#</a> [19:13] &lt;AryehGregor&gt; Only IE9 ever passes my new shiny assert_interface().</li>
 <li id="l-663" _a="AryehGregor"><a href="#l-663">#</a> [19:13] &lt;AryehGregor&gt; Go IE.</li>
 <li id="l-664" _a="AryehGregor"><a href="#l-664">#</a> [19:14] &lt;AryehGregor&gt; Gecko doesn't have Node.prototype.constructor.</li>
 <li id="l-665" _a="AryehGregor"><a href="#l-665">#</a> [19:15] &lt;AryehGregor&gt; Node.prototype is enumerable and configurable in Opera, and should be neither.</li>
 <li id="l-666"><a href="#l-666">#</a> [19:15] * Joins: plutoniix (~plutoniix@ppp-124-120-69-100.revip2.asianet.co.th)</li>
 <li id="l-667" _a="AryehGregor"><a href="#l-667">#</a> [19:15] &lt;AryehGregor&gt; In WebKit, window.Node is not writable but is enumerable, while it should be writable but not enumerable.  And String(Node) is wrong.</li>
 <li id="l-668" _a="Ms2ger" _r="AryehGregor"><a href="#l-668">#</a> [19:18] &lt;Ms2ger&gt; AryehGregor, can you also try it on NodeList?</li>
 <li id="l-669" _a="AryehGregor"><a href="#l-669">#</a> [19:19] * AryehGregor tries</li>
 <li id="l-670" _a="AryehGregor" _r="Ms2ger"><a href="#l-670">#</a> [19:20] &lt;AryehGregor&gt; Ms2ger, works.</li>
 <li id="l-671" _a="AryehGregor"><a href="#l-671">#</a> [19:20] &lt;AryehGregor&gt; I mean, it passes everything.</li>
 <li id="l-672" _a="Ms2ger"><a href="#l-672">#</a> [19:20] &lt;Ms2ger&gt; Excellent</li>
 <li id="l-673"><a href="#l-673">#</a> [19:41] * Joins: andyg (~andyg@101.170.200.168)</li>
 <li id="l-674"><a href="#l-674">#</a> [19:46] * Quits: andyg (~andyg@101.170.200.168) (Quit: andyg)</li>
 <li id="l-675"><a href="#l-675">#</a> [19:47] * Joins: dbaron (~dbaron@60.234.54.74)</li>
 <li id="l-676"><a href="#l-676">#</a> [19:56] * Quits: mpt (~mpt@canonical/mpt) (Ping timeout: 260 seconds)</li>
 <li id="l-677" _a="AryehGregor"><a href="#l-677">#</a> [20:02] &lt;AryehGregor&gt; What's an example of a static operation that's actually used in real life?  I don't find any in DOM4, and the only one I find in HTML is URL.createObjectURL.</li>
 <li id="l-678" _a="Ms2ger"><a href="#l-678">#</a> [20:07] &lt;Ms2ger&gt; IDB</li>
 <li id="l-679" _a="Ms2ger"><a href="#l-679">#</a> [20:07] &lt;Ms2ger&gt; Those are the only two</li>
 <li id="l-680"><a href="#l-680">#</a> [20:07] * Quits: cpearce (~chatzilla@ip-118-90-36-154.xdsl.xnet.co.nz) (Ping timeout: 260 seconds)</li>
 <li id="l-681"><a href="#l-681">#</a> [20:07] * Quits: FlorianX (~Dimitri@p4FCF7A7E.dip.t-dialin.net) (Quit: Leaving.)</li>
 <li id="l-682" _a="AryehGregor" _r="jgraham"><a href="#l-682">#</a> [20:11] &lt;AryehGregor&gt; jgraham, I submitted another testharness.js patch: <a href="http://www.w3.org/Bugs/Public/show_bug.cgi?id=14814">http://www.w3.org/Bugs/Public/show_bug.cgi?id=14814</a></li>
 <li id="l-683"><a href="#l-683">#</a> [20:24] * Joins: foolip__ (~philip@83.218.67.122)</li>
 <li id="l-684"><a href="#l-684">#</a> [20:25] * Quits: foolip (~philip@83.218.67.122) (Ping timeout: 260 seconds)</li>
 <li id="l-685" _a="heycam|away"><a href="#l-685">#</a> [20:37] * heycam|away is now known as heycam</li>
 <li id="l-686" _a="heycam" _r="AryehGregor"><a href="#l-686">#</a> [20:38] &lt;heycam&gt; AryehGregor, also you can test Object.getOwnPropertyDescriptor(obj, propname).configurable</li>
 <li id="l-687" _a="Ms2ger" _r="heycam"><a href="#l-687">#</a> [20:39] &lt;Ms2ger&gt; heycam, yes, that's what _bga said</li>
 <li id="l-688" _a="heycam" _r="Ms2ger"><a href="#l-688">#</a> [20:40] &lt;heycam&gt; Ms2ger, I'd have realised that if I'd read just one line further, I suppose :)</li>
 <li id="l-689" _a="Ms2ger"><a href="#l-689">#</a> [20:40] &lt;Ms2ger&gt; :)</li>
 <li id="l-690"><a href="#l-690">#</a> [20:41] * Joins: scor (~scor@drupal.org/user/52142/view)</li>
 <li id="l-691"><a href="#l-691">#</a> [20:43] * Joins: bga_ (~bga@ppp78-37-195-80.pppoe.avangarddsl.ru)</li>
 <li id="l-692"><a href="#l-692">#</a> [20:45] * Quits: _bga (~bga@95-55-61-178.dynamic.avangarddsl.ru) (Ping timeout: 260 seconds)</li>
 <li id="l-693"><a href="#l-693">#</a> [21:13] * Quits: smaug____ (~chatzilla@cs181151161.pp.htv.fi) (Ping timeout: 258 seconds)</li>
 <li id="l-694"><a href="#l-694">#</a> [21:33] * Quits: scor (~scor@drupal.org/user/52142/view) (Quit: scor)</li>
 <li id="l-695"><a href="#l-695">#</a> [21:37] * Joins: jdaggett (~jdaggett@60.234.54.74)</li>
 <li id="l-696"><a href="#l-696">#</a> [22:08] * Joins: smaug____ (~chatzilla@GZYYYMKCCCLXXVI.gprs.sl-laajakaista.fi)</li>
 <li id="l-697"><a href="#l-697">#</a> [22:14] * Joins: jdaggett_ (~jdaggett@60.234.54.74)</li>
 <li id="l-698"><a href="#l-698">#</a> [22:14] * Quits: jdaggett (~jdaggett@60.234.54.74) (Read error: Connection reset by peer)</li>
 <li id="l-699" _a="jdaggett_"><a href="#l-699">#</a> [22:14] * jdaggett_ is now known as jdaggett</li>
 <li id="l-700"><a href="#l-700">#</a> [22:32] * Quits: tomasf (~tom@c-b7dbe555.024-204-6c6b7012.cust.bredbandsbolaget.se) (Quit: tomasf)</li>
 <li id="l-701"><a href="#l-701">#</a> [22:32] * Joins: tomasf (~tom@c-b7dbe555.024-204-6c6b7012.cust.bredbandsbolaget.se)</li>
 <li id="l-702"><a href="#l-702">#</a> [22:32] * Quits: erlehmann (~erlehmann@82.113.121.67) (Quit: Ex-Chat)</li>
 <li id="l-703"><a href="#l-703">#</a> [22:36] * Quits: tomasf (~tom@c-b7dbe555.024-204-6c6b7012.cust.bredbandsbolaget.se) (Ping timeout: 240 seconds)</li>
 <li id="l-704"><a href="#l-704">#</a> [22:40] * Joins: scor (~scor@drupal.org/user/52142/view)</li>
 <li id="l-705" _a="gsnedders" _r="AryehGregor"><a href="#l-705">#</a> [22:45] &lt;gsnedders&gt; AryehGregor: getOPD often returns bogus things for host objects in general.</li>
 <li id="l-706" _a="Ms2ger" _r="Well"><a href="#l-706">#</a> [22:46] &lt;Ms2ger&gt; Well, then that's an excellent thing to test</li>
 <li id="l-707" _a="gsnedders"><a href="#l-707">#</a> [22:46] &lt;gsnedders&gt; A WebIDL mapping issue rather than a per-interface issue in general, though</li>
 <li id="l-708" _a="gsnedders"><a href="#l-708">#</a> [22:46] &lt;gsnedders&gt; And I'm not sure WebIDL actually requires it to be correct</li>
 <li id="l-709" _a="heycam"><a href="#l-709">#</a> [22:47] &lt;heycam&gt; hopefully it does</li>
 <li id="l-710"><a href="#l-710">#</a> [22:47] * Joins: rniwa (~rniwa@70-89-66-218-ca.sfba.hfc.comcastbusiness.net)</li>
 <li id="l-711" _a="gsnedders"><a href="#l-711">#</a> [22:48] &lt;gsnedders&gt; I have some draft email somewhere about things that WebIDL fails to specify.</li>
 <li id="l-712" _a="gsnedders"><a href="#l-712">#</a> [22:48] &lt;gsnedders&gt; I know what your intention is, heycam :)</li>
 <li id="l-713" _a="Ms2ger"><a href="#l-713">#</a> [22:48] &lt;Ms2ger&gt; Send it, dammit :)</li>
 <li id="l-714" _a="heycam" _r="gsnedders"><a href="#l-714">#</a> [22:48] &lt;heycam&gt; gsnedders, yes please send :)</li>
 <li id="l-715"><a href="#l-715">#</a> [22:49] * Quits: scor (~scor@drupal.org/user/52142/view) (Quit: scor)</li>
 <li id="l-716" _a="gsnedders" _r="Oh"><a href="#l-716">#</a> [22:51] &lt;gsnedders&gt; Oh, no, it's fixed now anyway</li>
 <li id="l-717" _a="Ms2ger"><a href="#l-717">#</a> [22:52] &lt;Ms2ger&gt; :)</li>
 <li id="l-718" _a="heycam"><a href="#l-718">#</a> [22:54] &lt;heycam&gt; yay</li>
 <li id="l-719"><a href="#l-719">#</a> [23:06] * Joins: mpt (~mpt@canonical/mpt)</li>
 <li id="l-720"><a href="#l-720">#</a> [23:07] * Quits: astearns (~anonymous@c-50-132-63-33.hsd1.wa.comcast.net) (Quit: astearns)</li>
 <li id="l-721"><a href="#l-721">#</a> [23:09] * Quits: Maurice (copyman@5ED573FA.cm-7-6b.dynamic.ziggo.nl)</li>
 <li id="l-722"><a href="#l-722">#</a> [23:09] * Quits: smaug____ (~chatzilla@GZYYYMKCCCLXXVI.gprs.sl-laajakaista.fi) (Ping timeout: 276 seconds)</li>
 <li id="l-723"><a href="#l-723">#</a> [23:14] * Quits: nw` (eero@heaven.unlink.org) (Ping timeout: 245 seconds)</li>
 <li id="l-724"><a href="#l-724">#</a> [23:16] * Joins: nw` (eero@heaven.unlink.org)</li>
 <li id="l-725"><a href="#l-725">#</a> [23:26] * Quits: jdaggett (~jdaggett@60.234.54.74) (Quit: jdaggett)</li>
 <li id="l-726"><a href="#l-726">#</a> [23:28] * Joins: jdaggett (~jdaggett@60.234.54.74)</li>
 <li id="l-727"><a href="#l-727">#</a> [23:29] * Joins: roc (~chatzilla@60.234.54.74)</li>
 <li id="l-728" _a="annevk"><a href="#l-728">#</a> [23:29] &lt;annevk&gt; heh</li>
 <li id="l-729" _a="annevk"><a href="#l-729">#</a> [23:30] &lt;annevk&gt; HTML5 new featured by WSJ: <a href="http://online.wsj.com/article/SB10001424052970203537304577030033160849296.html">http://online.wsj.com/article/SB10001424052970203537304577030033160849296.html</a></li>
 <li id="l-730" _a="Ms2ger"><a href="#l-730">#</a> [23:30] &lt;Ms2ger&gt; &quot;Implementations that use ECMAScript to implement the APIs defined in this specification must implement them in a manner consistent with the ECMAScript Bindings defined in the Web IDL specification&quot;</li>
 <li id="l-731" _a="Ms2ger"><a href="#l-731">#</a> [23:30] &lt;Ms2ger&gt; Does that mean C++ implementations are free to do whatever?</li>
 <li id="l-732" _a="annevk"><a href="#l-732">#</a> [23:30] &lt;annevk&gt; I hope so</li>
 <li id="l-733" _a="heycam" _r="&quot;implement&quot"><a href="#l-733">#</a> [23:31] &lt;heycam&gt; &quot;implement&quot; there is a bit iffy</li>
 <li id="l-734" _a="heycam"><a href="#l-734">#</a> [23:31] &lt;heycam&gt; it's not the implementation of the DOM needing to be js</li>
 <li id="l-735" _a="annevk"><a href="#l-735">#</a> [23:31] &lt;annevk&gt; we should definitely not make requirements on other languages I think</li>
 <li id="l-736" _a="heycam"><a href="#l-736">#</a> [23:31] &lt;heycam&gt; it should be saying if the APIs are exposed to JS</li>
 <li id="l-737" _a="heycam"><a href="#l-737">#</a> [23:31] &lt;heycam&gt; or something</li>
 <li id="l-738" _a="annevk" _r="heycam"><a href="#l-738">#</a> [23:32] &lt;annevk&gt; heycam: any chance we can get unions in?</li>
 <li id="l-739" _a="annevk" _r="heycam"><a href="#l-739">#</a> [23:32] &lt;annevk&gt; heycam: if not, I think I'm just going to use them when I get back to editing DOM4</li>
 <li id="l-740" _a="heycam"><a href="#l-740">#</a> [23:32] &lt;heycam&gt; ha</li>
 <li id="l-741" _a="annevk" _r="heycam"><a href="#l-741">#</a> [23:32] &lt;annevk&gt; heycam: much like Hixie's has been using non-existing features</li>
 <li id="l-742" _a="heycam"><a href="#l-742">#</a> [23:32] &lt;heycam&gt; I think they make sense</li>
 <li id="l-743" _a="heycam"><a href="#l-743">#</a> [23:33] &lt;heycam&gt; dunno if it is worth trying to drop all the overload stuff just in favour of unions</li>
 <li id="l-744" _a="Ms2ger" _r="Btw"><a href="#l-744">#</a> [23:33] &lt;Ms2ger&gt; Btw, cabal, please review the mouse lock spec</li>
 <li id="l-745" _a="heycam"><a href="#l-745">#</a> [23:33] &lt;heycam&gt; might be possible</li>
 <li id="l-746" _a="annevk" _r="Ms2ger"><a href="#l-746">#</a> [23:33] &lt;annevk&gt; Ms2ger: place a pointer in</li>
 <li id="l-747" _a="annevk"><a href="#l-747">#</a> [23:33] &lt;annevk&gt; though I'm going to try to sleep shortly</li>
 <li id="l-748" _a="Ms2ger"><a href="#l-748">#</a> [23:33] &lt;Ms2ger&gt; <a href="http://dvcs.w3.org/hg/webevents/raw-file/default/mouse-lock.html">http://dvcs.w3.org/hg/webevents/raw-file/default/mouse-lock.html</a></li>
 <li id="l-749" _a="annevk"><a href="#l-749">#</a> [23:33] &lt;annevk&gt; still adjusting after the flight :(</li>
 <li id="l-750" _a="annevk" _r="heycam"><a href="#l-750">#</a> [23:34] &lt;annevk&gt; heycam: there's definitely some overlap</li>
 <li id="l-751" _a="Ms2ger" _r="Hmm"><a href="#l-751">#</a> [23:34] &lt;Ms2ger&gt; Hmm, sleeping...</li>
 <li id="l-752" _a="annevk" _r="heycam"><a href="#l-752">#</a> [23:34] &lt;annevk&gt; heycam: though the append() stuff you cannot really do with overloading because of the varargs</li>
 <li id="l-753" _a="annevk"><a href="#l-753">#</a> [23:35] &lt;annevk&gt; allowing append(Comment, Text) does not work with append(Comment...); append(Text...) afaict</li>
 <li id="l-754" _a="annevk"><a href="#l-754">#</a> [23:36] &lt;annevk&gt; via this new &quot;Activity&quot; feature of twitter I find out that people &quot;favorite&quot; tweets a lot</li>
 <li id="l-755" _a="heycam" _r="annevk"><a href="#l-755">#</a> [23:36] &lt;heycam&gt; annevk, yeah that wouldn't allow append(aComment, aComment, aText, aComment)</li>
 <li id="l-756"><a href="#l-756">#</a> [23:44] * Joins: tantek (~tantek@208-90-212-140.PUBLIC.monkeybrains.net)</li>
 <li id="l-757"><a href="#l-757">#</a> [23:44] * Quits: rniwa (~rniwa@70-89-66-218-ca.sfba.hfc.comcastbusiness.net) (Quit: rniwa)</li>
 <li id="l-758"><a href="#l-758">#</a> [23:44] * Joins: shepazu (~shepazu@12.1.203.2)</li>
 <li id="l-759"><a href="#l-759">#</a> [23:54] * Quits: mpt (~mpt@canonical/mpt) (Ping timeout: 252 seconds)</li>
 <li id="l-760"><a href="#l-760">#</a> [23:54] * Quits: svl (~me@ip565744a7.direct-adsl.nl) (Quit: And back he spurred like a madman, shrieking a curse to the sky.)</li>
 <li id="l-761"><a href="#l-761">#</a> [23:58] * Joins: astearns (~anonymous@c-50-132-63-33.hsd1.wa.comcast.net)</li>
 <li id="l-762"><a href="#l-762">#</a> Session Close: Mon Nov 14 00:00:01 2011</li>
</ol>
<p id="end">The end :)
